% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{add_arrows}
\alias{add_arrows}
\title{Add arrow descriptions to a paths data frame}
\usage{
add_arrows(
  .paths,
  graph_params,
  graph_df_colnames,
  rebound_segments = ReboundTools::rebound_segments
)
}
\arguments{
\item{.paths}{The data frame to which arrows should be added.}

\item{graph_params}{See \code{ReboundTools::path_graph_params}.}

\item{graph_df_colnames}{See \code{ReboundTools::graph_df_colnames}.}

\item{rebound_segments}{See \code{ReboundTools::rebound_segments}.}
}
\value{
A version of \code{.paths} with a column for arrow descriptions.
}
\description{
Add arrow descriptions to a paths data frame
}
