% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{add_iso}
\alias{add_iso}
\title{Add an iso line description to a data frame}
\usage{
add_iso(
  .DF = NULL,
  indexed = FALSE,
  meta,
  graph_type,
  iso_name,
  colour = ReboundTools::graph_colours$grid,
  size = 0.5,
  linetype = "solid",
  x_orig,
  y_orig,
  x,
  y
)
}
\arguments{
\item{.DF}{}

\item{indexed}{}

\item{meta}{}

\item{graph_type}{}

\item{iso_name}{}

\item{colour}{}

\item{size}{}

\item{linetype}{}

\item{x_orig}{}

\item{y_orig}{}

\item{x}{}

\item{y}{}
}
\value{

}
\description{
Add an iso line description to a data frame
}
