% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{add_prefs_budget_line}
\alias{add_prefs_budget_line}
\title{Title}
\usage{
add_prefs_budget_line(
  .DF = NULL,
  meta,
  graph_type = ReboundTools::graph_types$preferences,
  line_name,
  colour = ReboundTools::graph_colours$grid,
  size = 0.5,
  linetype = "solid",
  slope,
  intercept
)
}
\arguments{
\item{.DF}{}

\item{meta}{}

\item{graph_type}{}

\item{line_name}{}

\item{colour}{}

\item{size}{}

\item{linetype}{}

\item{slope}{}

\item{intercept}{}
}
\value{

}
\description{
Title
}
