% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{add_segment}
\alias{add_segment}
\title{Add a line segment to a data frame}
\usage{
add_segment(
  .DF = NULL,
  indexed,
  meta,
  graph_type,
  segment_name,
  colour = "black",
  size = 1,
  linetype = "solid",
  x_orig,
  y_orig,
  x,
  y,
  xend,
  yend
)
}
\arguments{
\item{.DF}{}

\item{indexed}{}

\item{meta}{}

\item{graph_type}{}

\item{segment_name}{}

\item{colour}{}

\item{size}{}

\item{x_orig}{}

\item{y_orig}{}

\item{x}{}

\item{y}{}

\item{xend}{}

\item{yend}{}

\item{arrow}{}
}
\value{
A version of \code{.DF} with line segments added as the final row.
}
\description{
Add a line segment to a data frame
}
