% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bar_vars}
\alias{bar_vars}
\title{EEU bar data}
\format{
A string list with 12 entries.
\describe{
\item{eta_bar}{Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], exactly \code{eta_hat}.}
\item{p_s_bar}{The energy service price after the income effect [$/service], exactly \code{p_s_hat}.}
\item{C_dot_cap_bar}{The capital expenditure rate after the income effect [$/year], exactly \code{C_dot_cap_hat}.}
\item{C_dot_md_bar}{The maintenance and disposal expenditure rate after the income effect [$/year], exactly \code{C_dot_md_hat}.}
\item{E_dot_emb_bar}{The embodied energy rate after the income effect [MJ/year], exactly \code{E_dot_emb_hat}.}
\item{M_dot_bar}{Real income after the income effect [MJ/year], exactly \code{M_dot_hat}.}
\item{q_dot_s_bar}{The rate of energy service consumption after the income effect\ [service/year], calculated by \code{(1 + N_dot_hat/M_dot_hat_prime)^(e_qs_M)}.}
\item{C_dot_o_bar}{The rate of other goods expenditures after the income effect [$/year], calculated by \code{(1 + N_dot_hat/M_dot_hat_prime)^(e_qo_M)}.}
\item{N_dot_bar}{The freed cash rate after the income effect [$/year], exactly \code{0}.}
}

[service/year]: R:service/year\%5C
}
\usage{
bar_vars
}
\description{
This is the list of the derived variables at the bar stage (after income effect) of a rebound analysis.
}
\examples{
bar_vars
}
\keyword{datasets}
