% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_bar}
\alias{calc_bar}
\title{Calculate energy rebound data at the bar stage}
\usage{
calc_bar(
  .hat_data = NULL,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  C_dot_md_bar = ReboundTools::bar_vars$C_dot_md_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar
)
}
\arguments{
\item{.hat_data}{}
}
\value{
A list or data frame of derived rebound values for the bar stage (after the income effect).
}
\description{
This function calculates energy rebound information for the bar
stage (immediately after the income effect).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat() \%>\% 
  calc_bar()
}
