% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_Deltas}
\alias{calc_Deltas}
\title{Calculate differences between stages}
\usage{
calc_Deltas(
  .tilde_data = NULL,
  key_analysis_vars = ReboundTools::key_analysis_vars,
  rebound_stages = ReboundTools::rebound_stages,
  .all_vars = "all_vars"
)
}
\arguments{
\item{.tilde_data}{A data frame containing rebound calculations, original data,
star data, hat data, and tilde data,
likely calculated by \code{calc_tilde()}.}

\item{key_analysis_vars}{See \code{ReboundTools::key_analysis_vars}.}

\item{rebound_stages}{See \code{ReboundTools::rebound_stages}.}

\item{.all_vars}{A column names used internally for all variables.
Default is ".all_vars".}
}
\value{
\code{.tilde_data} with additional columns containing all possible difference terms.
}
\description{
Calculate differences between stages
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat() \%>\% 
  calc_bar() \%>\% 
  calc_tilde() \%>\% 
  calc_Deltas()
}
