% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_orig}
\alias{calc_orig}
\title{Calculate energy rebound derived data}
\usage{
calc_orig(
  .eeu_data = NULL,
  eta_orig_engr_units = ReboundTools::eeu_base_data$eta_orig_engr_units,
  MJ_engr_unit = ReboundTools::eeu_base_data$MJ_engr_unit,
  q_dot_s_orig = ReboundTools::eeu_base_data$q_dot_s_orig,
  eta_orig = ReboundTools::eeu_derived_data$eta_orig,
  E_dot_s_orig = ReboundTools::eeu_derived_data$E_dot_s_orig
)
}
\arguments{
\item{eta_orig, eta_tilde}{See \code{ReboundTools::eeu_base_data}.}

\item{.rebound_data}{An optional data frame containing EEU base data. See \code{ReboundTools::eeu_base_data}.}
}
\value{
A list or data frame of derived rebound values.
}
\description{
Calculate energy rebound derived data
}
\examples{
load_eeu_data() \%>\% 
  calc_orig()
}
