% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_rebound_data}
\alias{calc_rebound_data}
\title{Calculate energy rebound derived data}
\usage{
calc_rebound_data(
  .eeu_data = NULL,
  eta_orig = ReboundTools::eeu_base_data$eta_orig,
  eta_tilde = ReboundTools::eeu_base_data$eta_tilde,
  eta_ratio = ReboundTools::eeu_derived_data$eta_ratio
)
}
\arguments{
\item{eta_orig, eta_tilde}{See \code{ReboundTools::eeu_base_data}.}

\item{.rebound_data}{An optional data frame containing EEU base data. See \code{ReboundTools::eeu_base_data}.}
}
\value{
A list or data frame of derived rebound values.
}
\description{
Calculate energy rebound derived data
}
