% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_tilde}
\alias{calc_tilde}
\title{Calculate energy rebound data at the tilde stage}
\usage{
calc_tilde(
  .bar_data = NULL,
  tol = 1e-06,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  C_dot_md_bar = ReboundTools::bar_vars$C_dot_md_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar,
  eta_tilde = ReboundTools::tilde_vars$eta_tilde,
  p_s_tilde = ReboundTools::tilde_vars$p_s_tilde,
  C_dot_cap_tilde = ReboundTools::tilde_vars$C_dot_cap_tilde,
  C_dot_md_tilde = ReboundTools::tilde_vars$C_dot_md_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  M_dot_tilde = ReboundTools::tilde_vars$M_dot_tilde,
  q_dot_s_tilde = ReboundTools::tilde_vars$q_dot_s_tilde,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  C_dot_s_tilde = ReboundTools::tilde_vars$C_dot_s_tilde,
  C_dot_o_tilde = ReboundTools::tilde_vars$C_dot_o_tilde,
  N_dot_tilde = ReboundTools::tilde_vars$N_dot_tilde
)
}
\arguments{
\item{.bar_data}{}
}
\value{
A list or data frame of derived rebound values for the bar stage (after the income effect).
}
\description{
This function calculates energy rebound information for the tilde
stage (immediately after the productivity effect).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat() \%>\% 
  calc_bar() \%>\% 
  calc_tilde()
}
