% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_k}
\alias{calibrate_k}
\title{Calibrate values for macro parameter k}
\usage{
calibrate_k(
  .rebound_data,
  Re_tot_target,
  case_colname = ReboundTools::eeu_base_params$case,
  I_E_colname = ReboundTools::eeu_base_params$I_E,
  S_dot_dev_colname = ReboundTools::star_vars$S_dot_dev,
  N_dot_star_colname = ReboundTools::star_vars$N_dot_star,
  Re_macro_colname = ReboundTools::rebound_terms$Re_macro,
  Re_tot_colname = ReboundTools::rebound_terms$Re_tot
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound data, probably created by \code{rebound_analysis()}.}

\item{Re_tot_target}{The target value for total rebound.
A number or a vector of numbers of same length as the number of observations
in \code{.rebound_data}.}

\item{case_colname, I_E_colname}{See \code{ReboundTools::eeu_base_params}.}

\item{S_dot_dev_colname}{See \code{ReboundTools::star_vars}.}

\item{N_dot_star_colname}{See \code{ReboundTools::star_vars}.}

\item{Re_macro_colname, Re_tot_colname}{See \code{ReboundTools::rebound_terms}.}
}
\value{
A number or vector of numbers of $k$ values which make total rebound match \code{target_Re_tot}.
}
\description{
The macro parameter \code{k} is the least-well-known parameter in rebound analysis.
\code{k} links micro and macro rebound effects, and can be calibrated if
macro rebound (\code{Re_macro}) is considered to be a residual and \code{k} selected
such that a desired total rebound value (\code{target_Re_tot}) is obtained.
}
\details{
The calibrated value of \code{k} is calculated by
\verb{S_dot_dev / (N_dot_star * I_E) * [Re_macro + (Re_tot_target - Re_tot)]}.

Names of return values are taken from the case column in \code{.rebound_data}.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  calibrate_k(0.63)
}
