% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames_latex_key_analysis_vars}
\alias{cnames_latex_key_analysis_vars}
\title{Column names for LaTeX key analysis variables}
\format{
A list with 2 columns.
\describe{
\item{var_name}{The name for the column of variable names.}
\item{latex_var_name}{The name of the column of LaTeX variable names.}
}
}
\usage{
cnames_latex_key_analysis_vars
}
\description{
A list of column names for \code{latex_key_analysis_vars}.
}
\examples{
cnames_latex_key_analysis_vars
}
\keyword{datasets}
