% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{cost_graph}
\alias{cost_graph}
\title{Title}
\usage{
cost_graph(.rebound_data)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
Title
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  cost_graph()
}
