% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{cost_paths}
\alias{cost_paths}
\title{Title}
\usage{
cost_paths(
  .rebound_data,
  indexed = FALSE,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  G_dot_colour = ReboundTools::graph_colours$empl,
  G_dot_size = 1,
  Delta_C_dot_cap_star = ReboundTools::Delta_vars$Delta_C_dot_cap_star,
  Delta_C_dot_cap_star_colour = "black",
  Delta_C_dot_cap_star_size = 0.5,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_C_dot_md_star_colour = ReboundTools::graph_colours$empl,
  Delta_C_dot_md_star_size = 1,
  Delta_C_dot_s_hat = ReboundTools::Delta_vars$Delta_C_dot_s_hat,
  Delta_C_dot_s_hat_colour = ReboundTools::graph_colours$sub,
  Delta_C_dot_s_hat_size = 1,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_C_dot_o_hat_colour = ReboundTools::graph_colours$sub,
  Delta_C_dot_o_hat_size = 1,
  Delta_C_dot_s_bar = ReboundTools::Delta_vars$Delta_C_dot_s_bar,
  Delta_C_dot_s_bar_colour = ReboundTools::graph_colours$inc,
  Delta_C_dot_s_bar_size = 0.5,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  Delta_C_dot_o_bar_colour = ReboundTools::graph_colours$inc,
  Delta_C_dot_o_bar_size = 0.5,
  cost_type = ReboundTools::graph_types$cost
)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
Title
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  cost_paths()
}
