% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Delta_vars}
\alias{Delta_vars}
\title{Difference variables}
\format{
A string list with 68 entries.
\describe{
\item{Delta_eta_star}{Calculated as \code{eta_star - eta_orig}.}
\item{Delta_p_s_star}{Calculated as \code{p_s_star - p_s_orig}.}
\item{.}{Etc.}
\item{.}{Etc.}
\item{.}{Etc.}
\item{Delta_M_dot_tilde}{Calculated as \code{M_dot_tilde - M_dot_bar}.}
\item{Delta_N_dot_tilde}{Calculated as \code{N_dot_tilde - N_dot_bar}.}
}
}
\usage{
Delta_vars
}
\description{
This is the list of difference variables of the form \code{Delta_var_stage}, where
\code{var} is the variable name (from \code{ReboundTools::key_analysis_vars}) and
\code{stage} is the rebound stage being considered (from \code{ReboundTools::rebound_stages}).
}
\examples{
Delta_vars
}
\keyword{datasets}
