% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeu_base_params}
\alias{eeu_base_params}
\title{EEU base parameters}
\format{
A string list with 22 entries.
\describe{
\item{case}{A string to identify the case being analyzed, e.g., "Lamp".}
\item{original}{A string to identify the original device, e.g., "Incandescent".}
\item{upgrade}{A string to identify the upgraded device, e.g., "LED".}
\item{MJ_engr_unit}{A unit conversion factor: the number of MJ per engineering unit for the service efficiency. For example, if the service efficiency is given in miles/gallon, \code{MJ_engr_unit} should be 126.6 MJ/gallon. This unit conversion number is used in calculating the actual service efficiency.}
\item{I_E}{The energy intensity of the economy [MJ/$].}
\item{k}{The productivity effect factor [--].}
\item{p_E}{The price of energy [$/MJ].}
\item{eta_orig_engr_units}{The original (pre-EEU) energy service efficiency.  This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of \code{eta_orig_engr_units} is assumed to be the same as the denominator unit of \code{MJ_engr_unit}.}
\item{eta_tilde_engr_units}{The upgraded (post-EEU) energy service efficiency. This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of \code{eta_orig_engr_units} is assumed to be the same as the denominator unit of \code{MJ_engr_unit}.}
\item{e_qs_ps_UC}{The uncompensated ("UC") energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [--].}
\item{e_qs_M}{The income ("M") elasticity ("e") of energy service ("qs") consumption [--].}
\item{e_qo_M}{The income ("M") elasticity ("e") of other goods ("qo") consumption [--].}
\item{q_dot_s_orig}{The original (pre-EEU) consumption rate of the energy service. Example units are [miles/year] [lumen-hours/year].}
\item{M_dot_orig}{The disposable income rate, exclusive of taxes and savings [$/year].}
\item{C_cap_orig}{The net capital cost of the original device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].}
\item{t_orig}{The expected lifetime of the original device [year].}
\item{C_cap_star}{The net capital cost of the upgraded device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].}
\item{t_star}{The expected lifetime of the upgraded device [year].}
\item{C_dot_md_orig}{The original (pre-EEU) maintenance and disposal cost rate [$/year].}
\item{C_dot_md_star}{The upgraded (post-EEU) maintenance and disposal cost rate [$/year].}
\item{E_emb_orig}{The embodied energy of the original (pre-EEU) device [MJ].}
\item{E_emb_star}{The embodied energy of the upgraded (post-EEU) device [MJ].}
}
}
\usage{
eeu_base_params
}
\description{
This is the list of the required parameters for a rebound analysis.
These parameters must be supplied for a rebound analysis to commence.
The names in this list are the assumed names throughout this package,
but they can always be overridden by function arguments.
}
\details{
In the descriptions for each parameter, typical units are provided for
energy efficiency upgrades for cars and electric lamps.

Note the following meanings for the default parameter names:
\describe{
\item{orig}{The original device, prior to the EEU.}
\item{star}{After the emplacement effect and before the substitution effect.}
\item{hat}{After the substitution effect and before the income effect.}
\item{bar}{After the income effect and before the productivity effect.}
\item{tilde}{After the productivity effect.}
\item{I}{An economic intensity (per $).}
\item{p}{Price.}
\item{E}{Energy, typically final energy.}
\item{s}{Energy service.}
\item{o}{Other goods.}
\item{eta}{Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service.}
\item{eps}{Elasticity.}
\item{UC}{Uncompensated.}
\item{dot}{Signifies a rate, typically per year.}
\item{q}{Quantity of energy service or other goods consumed.}
\item{M}{Income.}
\item{C}{Cost.}
\item{t}{Expected device lifetime.}
}
}
\examples{
eeu_base_params
}
\keyword{datasets}
