% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeu_data_table}
\alias{eeu_data_table}
\title{EEU data table metadata}
\format{
A string list with 1 entries.
\describe{
\item{eeu_data_sheet}{The name of the sheet in an Excel file contining EEU data.}
}
}
\usage{
eeu_data_table
}
\description{
An EEU (energy efficiency upgrade) data table
contains user-supplied information needed to estimate energy rebound.
The \code{ReboundTools} package defines a standard nomenclature
for the columns in EEU data tables.
This object documents that nomenclature.
}
\examples{
eeu_data_table
}
\keyword{datasets}
