% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeu_derived_data}
\alias{eeu_derived_data}
\title{EEU derived data}
\format{
A string list with 22 entries.
\describe{
\item{eta_orig}{Energy service efficiency of the original (pre-EEU) device on a per-MJ basks [service/MJ].}
\item{eta_tilde}{Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ].}
\item{eta_ratio}{The ratio \code{eta_tilde/eta_orig} [--].}
\item{E_dot_s_orig}{\code{q_dot_s_orig / eta_orig}. [MJ/year].}
\item{S_dot_dev}{The expected device-level energy savings rate [MJ/year].}
}
}
\usage{
eeu_derived_data
}
\description{
This is the list of the derived parameters for a rebound analysis.
Each of these parameters is calculated data in \code{ReboundTools::eeu_base_data}.
}
\examples{
eeu_derived_data
}
\keyword{datasets}
