% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{energy_paths}
\alias{energy_paths}
\title{A data frame of energy rebound paths}
\usage{
energy_paths(
  .rebound_data,
  indexed = FALSE,
  k = ReboundTools::eeu_base_params$k,
  I_E = ReboundTools::eeu_base_params$I_E,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars$Delta_E_dot_emb_star,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars$Delta_E_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_E_dot_s_bar = ReboundTools::Delta_vars$Delta_E_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  S_dot_dev_colour = ReboundTools::graph_colours$empl,
  Delta_E_dot_emb_star_colour = ReboundTools::graph_colours$empl,
  Delta_C_dot_md_star_I_E_colour = "black",
  Delta_E_dot_s_hat_colour = ReboundTools::graph_colours$sub,
  Delta_C_dot_o_hat_I_E_colour = ReboundTools::graph_colours$sub,
  Delta_E_dot_s_bar_colour = ReboundTools::graph_colours$inc,
  Delta_C_dot_o_bar_I_E_colour = ReboundTools::graph_colours$inc,
  prod_colour = ReboundTools::graph_colours$prod,
  S_dot_dev_size = 0.1,
  Delta_E_dot_emb_star_size = 1,
  Delta_C_dot_md_star_I_E_size = 0.5,
  Delta_E_dot_s_hat_size = 1,
  Delta_C_dot_o_hat_I_E_size = 1,
  Delta_E_dot_s_bar_size = 0.5,
  Delta_C_dot_o_bar_I_E_size = 0.5,
  prod_size = 1,
  graph_type = ReboundTools::graph_types$energy
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound analysis results,
likely created by \code{rebound_analysis()}.}

\item{indexed}{A boolean telling whether the rebound path should be indexed to \code{1}
at its start.}

\item{k, I_E}{See \code{ReboundTools::eeu_base_params}.}

\item{E_dot_s_orig, E_dot_emb_orig, C_dot_md_orig, C_dot_o_orig}{See \code{ReboundTools::orig_vars}.}

\item{S_dot_dev}{See \code{ReboundTools::star_vars}.}

\item{N_dot_hat}{See \code{ReboundTools::hat_vars}.}

\item{Delta_E_dot_emb_star, Delta_C_dot_md_star, Delta_C_dot_o_hat, Delta_E_dot_s_bar, Delta_C_dot_o_bar}{See \code{ReboundTools::Delta_vars}.}

\item{S_dot_dev_colour, S_dot_dev_colour, Delta_E_dot_emb_star_colour, Delta_E_dot_s_hat_colour, Delta_C_dot_o_hat_I_E_colour, Delta_E_dot_s_bar_colour, Delta_C_dot_o_bar_I_E_colour, prod_colour}{See \code{ReboundTools::graph_colours}.}

\item{Delta_C_dot_md_star_I_E_colour}{Colour for the maintenance and disposal line. Default is "black".}

\item{S_dot_dev_size, Delta_E_dot_emb_star_size, Delta_C_dot_md_star_I_E_size, Delta_E_dot_s_hat_size, Delta_C_dot_o_hat_I_E_size, Delta_E_dot_s_bar_size, Delta_C_dot_o_bar_I_E_size, prod_size}{Line widths for energy rebound segments.}

\item{graph_type}{See \code{ReboundTools::graph_types}.}
}
\value{
A data frame with energy rebound path segments.
}
\description{
Make a data frame of segments
for an energy rebound graph.
Each stage of the rebound process is represented in the data frame.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  energy_paths()
}
