% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{energy_points}
\alias{energy_points}
\title{A data frame of energy rebound points}
\usage{
energy_points(
  .energy_paths,
  graph_params = ReboundTools::default_graph_params,
  rebound_stages = ReboundTools::rebound_stages,
  rebound_segments = ReboundTools::rebound_segments
)
}
\arguments{
\item{.energy_paths}{A data frame of energy paths,
likely created by \code{energy_paths()}.}

\item{graph_params}{See \code{ReboundTools::graph_params}.}

\item{rebound_stages}{See \code{ReboundTools::rebound_stages}.}

\item{rebound_segments}{See \code{ReboundTools::rebound_segments}.}
}
\value{
A data frame with energy rebound path segments.
}
\description{
Make a data frame of points between energy rebound effects
for an energy rebound graph.
Each stage of the rebound process is represented in the data frame.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  energy_paths() \%>\% 
  energy_points()
}
