% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{extract_meta}
\alias{extract_meta}
\title{Extract a metadata data frame from rebound data}
\usage{
extract_meta(
  .rebound_data,
  meta_cols = c(reference, case, original, upgrade),
  reference = ReboundTools::eeu_base_params$reference,
  case = ReboundTools::eeu_base_params$case,
  original = ReboundTools::eeu_base_params$original,
  upgrade = ReboundTools::eeu_base_params$upgrade
)
}
\arguments{
\item{.rebound_data}{The data frame from which metadata is to be extracted.}

\item{meta_cols}{A vector of column names for metadata columns. Default is \code{c(reference, case, original, upgrade)}.}

\item{reference, case, original, upgrade}{Names of metadata columns. See \code{ReboundTools::eeu_base_params}.}
}
\value{
A data frame containing only metadata columns.
}
\description{
Many functions in \code{ReboundTools} use the metadata from energy efficiency upgrade
parameters as a starting point for rebound paths or other graph creation.
This function provides a way to quickly extract that metadata.
}
\examples{
load_eeu_data() \%>\% 
  extract_meta()
}
