% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{extract_points}
\alias{extract_points}
\title{Extract points from a data frame of paths}
\usage{
extract_points(
  .paths,
  graph_params = ReboundTools::path_graph_params,
  rebound_stages = ReboundTools::rebound_stages,
  rebound_segments = ReboundTools::rebound_segments,
  graph_df_colnames = ReboundTools::graph_df_colnames
)
}
\arguments{
\item{.paths}{A data frame of rebound paths,
likely created by \code{energy_paths()}, \code{cost_paths()}, or \code{prefs_paths()}.}

\item{graph_params}{See \code{ReboundTools::graph_params}.}

\item{rebound_stages}{See \code{ReboundTools::rebound_stages}.}

\item{rebound_segments}{See \code{ReboundTools::rebound_segments}.}

\item{graph_df_colnames}{See \code{ReboundTools::graph_df_colnames}.}
}
\value{
A data frame containing points extracted from paths.
}
\description{
Make a data frame of points between energy rebound effects
for any type of rebound path graph.
Each available stage of the rebound process is represented in the data frame.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  energy_paths() \%>\% 
  extract_points()
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  cost_paths() \%>\% 
  extract_points()
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  prefs_paths() \%>\% 
  extract_points()
}
