% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indifference.R
\name{geom_seq}
\alias{geom_seq}
\title{Generate a geometric sequence}
\usage{
geom_seq(from, to, n)
}
\arguments{
\item{from}{The first value in the sequence.}

\item{to}{The final value in the sequence.}

\item{n}{The number of points to include in the sequence.}

\item{by}{The value by which each term in the sequence is multiplied to obtain the next term.}
}
\value{
A vector of numbers.
}
\description{
Generate a geometric sequence
}
\examples{
geom_seq(from = 0.01, to = 10, by = 1.1)
geom_seq(from = 10, to = 20, by = 0)
}
