% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_colours}
\alias{graph_colours}
\title{Graph colours}
\format{
A string list with 5 entries.
\describe{
\item{empl}{The colour for emplacement effects.}
\item{sub}{The colour for substitution effects.}
\item{inc}{The colour for income effects.}
\item{prod}{The colour for productivity effects.}
\item{grid}{The colour for grid decorations.}
}
}
\usage{
graph_colours
}
\description{
The list of graph colours in the \code{ReboundTools} package.
These are default colors for various graphical elements.
}
\examples{
graph_colours
}
\keyword{datasets}
