% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_types}
\alias{graph_types}
\title{Plane types}
\format{
A string list with 3 entries.
\describe{
\item{energy}{The energy plane graph type.}
\item{expenditure}{The expenditure plane graph type.}
\item{consumption}{The consumption plane graph type.}
}
}
\usage{
graph_types
}
\description{
The list of plane types in the \code{ReboundTools} package.
}
\examples{
graph_types
}
\keyword{datasets}
