% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{indifference_fun}
\alias{indifference_fun}
\title{An indifference curve}
\usage{
indifference_fun(
  qs_qs0,
  qs1_qs0,
  Co1_Co0,
  f_Cs,
  sigma,
  rho = (sigma - 1)/sigma
)
}
\arguments{
\item{qs_qs0}{The ratio \code{q_dot_s/q_dot_s_orig}. Sweeping this variable (x) gives
the indifference curve parameterized by the other arguments.}

\item{qs1_qs0}{The x coordinate of a point on this indifference curve.}

\item{Co1_Co0}{The y coordinate of a point on this indifference curve.}

\item{f_Cs}{The fraction of original spending on the energy service relative to the sum of energy service and other goods spending, calculated by \code{C_dot_s_orig / (C_dot_s_orig + C_dot_o_orig)}.}

\item{sigma}{The elasticity of substitution between the energy service and other goods.}

\item{rho}{The exponent in the CES utility function. Default is \code{(sigma-1)/sigma}.}
}
\value{
The value of \code{C_dot_o/C_dot_o_orig}, given values for remaining arguments.
}
\description{
This function gives points along an indifference curve in
(q_dot_s/q_dot_s_orig, C_dot_o/C_dot_o_orig) space.
The indifference curve assumes CES utility.
The equation of the indifference curve is
\verb{u_dot/u_dot_orig = [f_Cs*(q_dot_s/q_dot_s_orig)^rho + (1-f_Cs)*(C_dot_o/C_dot_o_orig)^rho]^(1/rho)}.
}
\examples{

}
