% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{iso_cost_lines}
\alias{iso_cost_lines}
\title{Create a data frame of iso-cost lines}
\usage{
iso_cost_lines(
  .rebound_data,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound information,
likely created by \code{rebound_analysis()}.}
}
\value{

}
\description{
This function creates a data frame of constant-energy or constant-cost lines
at various rebound stages.
The iso lines are described by slope and intercept.
}
