% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{iso_energy_lines}
\alias{iso_energy_lines}
\title{Title}
\usage{
iso_energy_lines(
  .rebound_data,
  indexed = FALSE,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  I_E = ReboundTools::eeu_base_params$I_E,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  C_dot_md_bar = ReboundTools::bar_vars$C_dot_md_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  C_dot_md_tilde = ReboundTools::tilde_vars$C_dot_md_tilde,
  C_dot_o_tilde = ReboundTools::tilde_vars$C_dot_o_tilde,
  k = ReboundTools::eeu_base_params$k,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  energy_type = ReboundTools::graph_types$energy,
  grid_colour = ReboundTools::graph_colours$grid,
  grid_size = 0.5,
  grid_linetype = "solid"
)
}
\arguments{
\item{.rebound_data}{}

\item{indexed}{}
}
\value{

}
\description{
Title
}
