% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latex_key_analysis_vars}
\alias{latex_key_analysis_vars}
\title{LaTeX key analysis variables}
\format{
A data frame with 3 columns.
\describe{
\item{description}{Short description of the variable in this row.}
\item{var_name}{The text name for key analysis variables.}
\item{latex_var_name}{The LaTeX key analysis variable name.}
}
}
\usage{
latex_key_analysis_vars
}
\description{
A data frame of key analysis variables in LaTeX formatting.
}
\examples{
latex_key_analysis_vars
}
\keyword{datasets}
