% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latex_rebound_stages}
\alias{latex_rebound_stages}
\title{LaTeX rebound stages}
\format{
A data frame with 2 columns.
\describe{
\item{stage}{The text name for the stage.}
\item{latex_stage_name}{The LaTeX name for the stage.}
}
}
\usage{
latex_rebound_stages
}
\description{
A data frame of rebound stages in LaTeX formatting.
}
\examples{
latex_rebound_stages
}
\keyword{datasets}
