% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latex_rebound_terms}
\alias{latex_rebound_terms}
\title{LaTeX version of rebound terms}
\format{
A string list with 15 entries.
\describe{
\item{Re_dempl}{Direct emplacement effect rebound, always 0.}
\item{Re_emb}{Indirect embodied energy effect rebound.}
\item{Re_cap}{Indirect capital expenditure effect rebound.}
\item{Re_md}{Indirect maintenance and disposal effect rebound.}
\item{Re_empl}{Emplacement effect rebound.}
\item{Re_dsub}{Direct substitution effect rebound.}
\item{Re_isub}{Indirect substitution effect rebound.}
\item{Re_sub}{Substitution effect rebound.}
\item{Re_dinc}{Direct income effect rebound.}
\item{Re_iinc}{Indirect income effect rebound.}
\item{Re_inc}{Income effect rebound.}
\item{Re_macro}{Indirect macro effect rebound.}
\item{Re_d}{Sum of all direct rebound effects.}
\item{Re_i}{Sum of all indirect rebound effects.}
\item{Re_tot}{Total rebound.}
}
}
\usage{
latex_rebound_terms
}
\description{
This is the list of the rebound terms in LaTeX form.
}
\examples{
latex_rebound_terms
}
\keyword{datasets}
