% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orig_vars}
\alias{orig_vars}
\title{EEU orig data}
\format{
A string list with 10 entries.
\describe{
\item{eta_orig}{Energy service efficiency of the original (pre-EEU) device on a per-MJ basis [service/MJ], calculated by \code{eta_orig_engr_units / MJ_engr_unit}.}
\item{E_dot_s_orig}{The final energy consumption rate of the original (pre-EEU) device [MJ/year], calculated by \code{q_dot_s_orig / eta_orig}.}
\item{C_dot_cap_orig}{The original (pre-EEU) capital cost rate [$/year], calculated by \code{C_cap_orig / t_orig}.}
\item{p_s_orig}{The original (pre-EEU) energy service price [$/service], calculated by \code{p_E / eta_orig}.}
\item{C_dot_s_orig}{The original (pre-EEU) rate of energy cost expenditures for the device [$/year], calculated by \code{p_E * E_dot_s_orig}.}
\item{C_dot_o_orig}{The original (pre-EEU) rate of expenditure on other goods [$/year], calculated by \code{M_dot_orig - C_dot_s_orig - C_dot_cap_orig - C_dot_md_orig}.}
\item{f_Cs_orig}{The original (pre-EEU) fraction of the energy and other budget spent on the energy service [--], calculated by \code{C_dot_s_orig / (C_dot_s_orig + C_dot_o_orig)}.}
\item{e_qs_ps}{The energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [--], calculated by \code{e_qs_ps_UC_val + f_Cs_orig_val*e_qs_M_val}.}
\item{e_qo_ps}{The energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [--], calculated by \code{f_Cs_orig_val*(f_Cs_orig_val + e_qs_ps_UC_val) / (f_Cs_orig_val - 1)}.}
\item{E_dot_emb_orig}{The original (pre-EEU) rate of embodied energy demand[MJ/year], calculated by \code{E_dot_emb_val / t_orig_val}.}
}
}
\usage{
orig_vars
}
\description{
This is the list of the derived variables at the original (pre-EEU) stage of a rebound analysis.
}
\examples{
orig_vars
}
\keyword{datasets}
