% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parametric_analysis_point_types}
\alias{parametric_analysis_point_types}
\title{Parametric analysis point types}
\format{
A string list with 3 entries.
\describe{
\item{point_type_colname}{The name of the column that contains parametric analysis point types.}
\item{orig}{The string for original points.}
\item{sweep}{The string for sweep points.}
}
}
\usage{
parametric_analysis_point_types
}
\description{
This is the list of parametric analysis point types.
}
\examples{
parametric_analysis_point_types
}
\keyword{datasets}
