% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{prefs_paths}
\alias{prefs_paths}
\title{Title}
\usage{
prefs_paths(
  .rebound_data,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_C_dot_o_hat_colour = ReboundTools::graph_colours$sub,
  Delta_C_dot_o_hat_size = 1,
  Delta_q_dot_s_hat = ReboundTools::Delta_vars$Delta_q_dot_s_hat,
  Delta_q_dot_s_hat_colour = ReboundTools::graph_colours$sub,
  Delta_q_dot_s_hat_size = 1,
  prefs_type = ReboundTools::graph_types$preferences
)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
Title
}
