% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_graphs}
\alias{rebound_graphs}
\title{Create rebound graphs}
\usage{
rebound_graphs(
  .rebound_data,
  indexed = FALSE,
  cases = .rebound_data[[case_colname]] \%>\% unique(),
  graph_types = ReboundTools::graph_types,
  grid_types = ReboundTools::graph_types,
  graph_params = ReboundTools::default_graph_params,
  case_colname = ReboundTools::eeu_base_params$case,
  graph_df_colnames = ReboundTools::graph_df_colnames
)
}
\arguments{
\item{.rebound_data}{Rebound data, likely loaded by \code{load_eeu_data()}.}
}
\value{
A \code{ggplot2} object with the graphs
}
\description{
Create rebound graphs
}
\examples{
load_eeu_data() \%>\% 
  rebound_graphs()
}
