% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_paths}
\alias{rebound_paths}
\title{Title}
\usage{
rebound_paths(
  .rebound_data,
  indexed = FALSE,
  I_E = ReboundTools::eeu_base_params$I_E,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  S_dot_dev_colour = "red",
  S_dot_dev_linetype = "dashed",
  energy_type = "Energy",
  cost_type = "Cost",
  prefs_type = "Preferences"
)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
Title
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  rebound_paths()
}
