% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rebound_results_table}
\alias{rebound_results_table}
\title{Build a rebound results table}
\usage{
rebound_results_table(
  .analysis_data,
  rebound_terms = ReboundTools::rebound_terms,
  ...
)
}
\arguments{
\item{.analysis_data}{Rebound analysis results. Probably the output of a call to \code{rebound_analysis()}.}

\item{rebound_terms}{See \code{ReboundTools::rebound_terms}.}

\item{...}{Arguments passed to \code{xtable::xtable()}, possibly
\code{label}, \code{caption}, \code{digits}, etc.}
}
\value{
A table of rebound analysis results suitable for printing.
}
\description{
Build a rebound results table
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  rebound_results_table()
}
