% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebound_segments}
\alias{rebound_segments}
\title{Rebound segment names}
\format{
A list frame with 9 entries
\describe{
\item{dempl}{Direct emplacement effects.}
\item{emb}{Embodied energy effects.}
\item{cap}{Capital cost effects.}
\item{md}{Maintenance and disposal effects.}
\item{dsub}{Direct substitution effects.}
\item{isub}{Indirect substitution effects.}
\item{dinc}{Direct income effects.}
\item{iinc}{Indirect income effects.}
\item{macro}{Macro effects.}
}
}
\usage{
rebound_segments
}
\description{
A list of rebound segment names used internally to the package.
}
\examples{
rebound_segments
}
\keyword{datasets}
