% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebound_stages}
\alias{rebound_stages}
\title{Rebound stages}
\format{
A string list with 5 entries.
\describe{
\item{orig}{The original (pre-EEU) stage.}
\item{star}{The upgraded condition (post-EEU), before any behavior changes.}
\item{hat}{After the substitution effect but before the income effect.}
\item{bar}{After the income effect but before the productivity effect.}
\item{tilde}{After the productivity effect.}
}
}
\usage{
rebound_stages
}
\description{
This is the list of rebound stages.
}
\examples{
rebound_stages
}
\keyword{datasets}
