% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebound_terms_agg}
\alias{rebound_terms_agg}
\title{Aggregate rebound terms}
\format{
A string list with 6 entries.
\describe{
\item{Re_empl}{Emplacement effect rebound.}
\item{Re_sub}{Substitution effect rebound.}
\item{Re_inc}{Income effect rebound.}
\item{Re_d}{Sum of all direct rebound effects.}
\item{Re_i}{Sum of all indirect rebound effects.}
\item{Re_tot}{Total rebound.}
}
}
\usage{
rebound_terms_agg
}
\description{
This is the list of the rebound terms that are aggregates of other rebound terms.
}
\examples{
rebound_terms_agg
}
\keyword{datasets}
