% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_terms_graph}
\alias{rebound_terms_graph}
\title{Title}
\usage{
rebound_terms_graph(
  .parametric_data = parametric_analysis(rebound_data, parameterization),
  rebound_data,
  parameterization,
  x_var,
  line_var = y_names_col,
  y_vals_col = "y_vals",
  y_names_col = "y_names",
  graph_params = ReboundTools::sens_graph_params,
  point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
  sweep_points = ReboundTools::parametric_analysis_point_types$sweep,
  orig_points = ReboundTools::parametric_analysis_point_types$orig
)
}
\arguments{
\item{.parametric_data}{}

\item{rebound_data}{}

\item{parameterization}{}

\item{x_var}{}

\item{line_var}{}

\item{y_vals_col}{}

\item{y_names_col}{}

\item{graph_params}{}

\item{point_type_colname}{}

\item{sweep_points}{}

\item{orig_points}{}
}
\value{

}
\description{
Title
}
