% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebound_units}
\alias{rebound_units}
\title{Information for rebound units}
\format{
A list with 10 entries.
\describe{
\item{energy_si}{The SI energy unit ("MJ").}
\item{time_unit}{The time unit ("yr").}
\item{inverse_time_unit}{Inverse time unit ("1/yr").}
\item{currency_unit}{The default currency unit ("$").}
\item{currency_unit_latex}{The currency unit in LaTeX format ("$" with escaping backslashes).}
\item{unitless}{The identifier for unitless variables ("-").}
\item{unitless_latex}{The identifier for unitless variables in LaTeX format ("--").}
\item{leading_delta_pattern}{A regex pattern that identifies a leading Delta ("^Delta_").}
\item{surround_left}{The left delimiter for units ("[").}
\item{surround_right}{The right delimiter for units ("]").}
}
}
\usage{
rebound_units
}
\description{
A list of names and constants for rebound units.
}
\examples{
rebound_units
}
\keyword{datasets}
