% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{rebound_var_units}
\alias{rebound_var_units}
\title{Units for variables}
\usage{
rebound_var_units(
  .var_name,
  service_unit,
  energy_engr_unit,
  escape_latex = FALSE,
  surround_left = ReboundTools::rebound_units$surround_left,
  surround_right = ReboundTools::rebound_units$surround_right,
  leading_delta_pattern = ReboundTools::rebound_units$leading_delta_pattern,
  energy_si = ReboundTools::rebound_units$energy_si,
  time_unit = ReboundTools::rebound_units$time_unit,
  currency = ReboundTools::rebound_units$currency_unit,
  currency_latex = ReboundTools::rebound_units$currency_unit_latex,
  unitless = ReboundTools::rebound_units$unitless,
  unitless_latex = ReboundTools::rebound_units$unitless_latex,
  service_unit_name = ReboundTools::eeu_base_params$service_unit,
  energy_engr_unit_name = ReboundTools::eeu_base_params$energy_engr_unit,
  energy_converter = ReboundTools::eeu_base_params$MJ_engr_unit,
  p_E_engr_units = "p_E_engr_units",
  p_E = "p_E",
  p_s = "p_s",
  q_dot_s = "q_dot_s",
  efficiency_engr_units = "eta_engr_units",
  efficiency = "eta",
  k = "k",
  I_E = "I_E",
  elasticities = "e",
  sigma = "sigma",
  time = "t_",
  cost_rate = "C_dot",
  cost = "C_",
  income_rate = "M_dot",
  income = "M",
  freed_cash_rate = "N_dot",
  freed_cash = "N",
  energy_rate = "E_dot",
  energy = "E",
  S_dot_dev = "S_dot_dev",
  G_dot = "G_dot",
  rebound = "Re_",
  f_Cs = "f_Cs"
)
}
\arguments{
\item{.var_name}{The variable name for which the unit is to be determined.}

\item{service_unit}{The energy service unit for this case (a string).}

\item{energy_engr_unit}{The engineering unit for energy (a string).}

\item{escape_latex}{A boolean that tells whether to encode the result as LaTeX output.}

\item{surround_left}{A string to add at the leading end of the result. Default is "[".}

\item{surround_right}{A string to add at the trailing end of the result. Default is "]".}

\item{leading_delta_pattern}{A regex pattern identifying a leading "Delta_". Default is "^Delta_".}

\item{energy_si}{}

\item{time_unit}{}

\item{currency}{}

\item{currency_latex}{}

\item{unitless}{}

\item{unitless_latex}{}

\item{service_unit_name}{}

\item{energy_engr_unit_name}{}

\item{energy_converter}{}

\item{p_E_engr_units}{}

\item{p_E}{}

\item{p_s}{}

\item{q_dot_s}{}

\item{efficiency_engr_units}{}

\item{efficiency}{}

\item{k}{}

\item{I_E}{}

\item{elasticities}{}

\item{sigma}{}

\item{time}{}

\item{cost_rate}{}

\item{cost}{}

\item{income_rate}{}

\item{income}{}

\item{freed_cash_rate}{}

\item{freed_cash}{}

\item{energy_rate}{}

\item{energy}{}

\item{S_dot_dev}{}

\item{G_dot}{}

\item{rebound}{}

\item{f_Cs}{}
}
\value{
A string for the units for \code{.var_name}.
}
\description{
Determine the unit for a variable in the rebound framework.
}
\details{
Information about the particular case is required,
thus \code{service_unit} and \code{energy_engr_unit} are arguments.

To surround the result with a string,
both \code{surround_left} and \code{surround_right} must be specified.
For example setting \code{surround_left = "["} and \code{surround_right = "]"} (the defaults)
gives "[unit]".

This function is vectorized.
}
\examples{
units("eta_engr_units_orig", service_unit = "lm-hr", energy_engr_unit = "kW-hr") 
units("p_s", service_unit = "lm-hr", energy_engr_unit = "kW-hr") 
units("p_s_orig", service_unit = "lm-hr", energy_engr_unit = "kW-hr") 
units("Delta_C_dot_o_hat", service_unit = "lm-hr", energy_engr_unit = "kW-hr") 
units(c("eta_engr_units_orig", "Delta_C_dot_o_hat"), service_unit = c("lm-hr", "lm-hr"), c(energy_engr_unit = "kW-hr"))
}
