% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{sample_eeu_data_path}
\alias{sample_eeu_data_path}
\title{Path to sample EEU data}
\usage{
sample_eeu_data_path()
}
\value{
A path to an example data file.
}
\description{
Sample energy efficiency upgrade (EEU) data gives examples and enables testing.
This function returns a string path to an Excel file containing
example data from the paper ---------------------.
Two energy efficiency upgrade examples are provided: a car and an electric lamp.
}
\details{
Users can provide their own EEU files in similar format to
make their own estimates of rebound effects.
}
\examples{
sample_eeu_data_path()
}
