% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stages_table_visibility_mask}
\alias{stages_table_visibility_mask}
\title{A visibility mask for the \code{stages_table()}}
\format{
A data frame with three columns, "Variable", "Stage", and "Visible".
\describe{
\item{.name}{The name of the variable in the \code{\link[=stages_table]{stages_table()}}.}
\item{.stage}{The name of the rebound in the \code{\link[=stages_table]{stages_table()}}.}
\item{Visible}{A boolean that tells whether this combination of \code{.name} and \code{.stage} should be visible.}
}
}
\usage{
stages_table_visibility_mask
}
\description{
This data frame contains a boolean column ("Visible")
that tells whether values are shown
in the \code{\link[=stages_table]{stages_table()}}.
}
\examples{
stages_table_visibility_mask
}
\keyword{datasets}
