% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{star_vars}
\alias{star_vars}
\title{EEU star data}
\format{
A string list with 11 entries.
\describe{
\item{eta_star}{Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], calculated by \code{eta_tilde_engr_units / MJ_engr_unit}.}
\item{eta_ratio}{The energy service efficiency ratio [--], calculated by \code{eta_star/eta_orig}.}
\item{S_dot_dev}{The expected device-level energy savings rate [MJ/year], calculated by \code{(eta_ratio - 1) * (1/eta_ratio) * E_dot_s_orig}.}
\item{G_dot}{The expected device-level energy gross cost savings rate [MJ/year], calculated by \code{p_E * S_dot_dev}.}
\item{p_s_star}{The upgraded (post-EEU) energy service price [$/service], calculated by \code{p_E / eta_star = p_E / eta_tilde}.}
\item{q_dot_s_star}{The upgraded (post-EEU) energy service consumption rate [service/year], same as \code{q_dot_s_orig}.}
\item{C_dot_cap_star}{The upgraded (post-EEU) capital cost rate [$/year], calculated by \code{C_cap_star / t_star}.}

\item{E_dot_s_star}{The final energy consumption rate of the original (pre-EEU) device [MJ/year], calculated by \code{q_dot_s_orig / eta_orig}.}
}
}
\usage{
star_vars
}
\description{
This is the list of the derived variables at the star (post-emplacement) stage of a rebound analysis.
}
\examples{
star_vars
}
\keyword{datasets}
