% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\title{Units for variables}
\usage{
units(
  .var_name,
  service_unit,
  energy_engr_unit,
  surround_left = "[",
  surround_right = "]",
  leading_delta = "^Delta_",
  trailing_stage = "_[^_]*$",
  energy_si = "MJ",
  efficiency_engy_units = "eta_engr_units",
  efficiency = "eta",
  cost_rate = "C_dot",
  currency = "$",
  time = "year"
)
}
\arguments{
\item{.var_name}{}

\item{service_unit}{}

\item{energy_engr_unit}{}

\item{leading_delta}{}

\item{trailing_stage}{}

\item{currency}{}

\item{time}{}
}
\value{

}
\description{
Units for variables
}
