% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{var_units}
\alias{var_units}
\title{Units for rebound variables}
\usage{
var_units(
  .rebound_data = load_eeu_data() \%>\% rebound_analysis(),
  case = ReboundTools::eeu_base_params$case,
  service_unit = ReboundTools::eeu_base_params$service_unit,
  energy_engr_unit = ReboundTools::eeu_base_params$energy_engr_unit
)
}
\arguments{
\item{.rebound_data}{A data frame, typically calculated by \code{rebound_analysis()}.
Default is \code{load_eeu_data() \%>\% rebound_analysis()}.}
}
\value{

}
\description{
This function calculates the units for rebound analysis variables.
}
\examples{
var_units()
}
