% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{add_indifference_curve}
\alias{add_indifference_curve}
\title{Title}
\usage{
add_indifference_curve(
  .DF = NULL,
  meta,
  graph_type = ReboundTools::graph_types$preferences,
  line_name,
  colour = ReboundTools::graph_colours$grid,
  size = 0.5,
  linetype = "solid",
  qs1_qs0,
  Co1_Co0,
  f_Cs_orig,
  sigma
)
}
\arguments{
\item{.DF}{}

\item{meta}{}

\item{graph_type}{}

\item{line_name}{}

\item{colour}{}

\item{size}{}

\item{linetype}{}

\item{qs1_qs0}{}

\item{Co1_Co0}{}

\item{f_Cs_orig}{}

\item{sigma}{}
}
\value{

}
\description{
Title
}
