% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{add_indifference_curve}
\alias{add_indifference_curve}
\title{Add an indifference curve to a data frame}
\usage{
add_indifference_curve(
  .DF = NULL,
  meta,
  graph_type = ReboundTools::graph_types$preferences,
  line_name,
  colour = ReboundTools::default_graph_params$prefs_indiff_grid_colour,
  size = ReboundTools::default_graph_params$prefs_indiff_grid_size,
  linetype = ReboundTools::default_graph_params$prefs_indiff_grid_linetype,
  qs1_qs0,
  Co1_Co0,
  f_Cs_orig,
  sigma,
  graph_df_colnames = ReboundTools::graph_df_colnames
)
}
\arguments{
\item{.DF}{A data frame that accumulates preferences curves.
When \code{NULL}, the default, a new data frame is created and returned.
When not \code{NULL}, rows for the curves are added to the bottom of \code{.DF}.}

\item{meta}{A data frame of metadata for the segment to be added.
This metadata data frame provides the left-most columns of the return value.}

\item{graph_type}{The graph type for the indifference curve.
Default is \code{ReboundTools::graph_types$preferences}.}

\item{line_name}{A name for this indifference curve.}

\item{colour}{The colour for this indifference curve.
Default is \code{ReboundTools::default_graph_params$prefs_indiff_curve_colour}.}

\item{size}{Line width. Default is \code{ReboundTools::default_graph_params$prefs_indiff_grid_size}.}

\item{linetype}{Line type. Default is \code{ReboundTools::default_graph_params$prefs_indiff_grid_linetype}.}

\item{qs1_qs0, Co1_Co0}{The (x,y) coordinates of a point on this indifference curve.}

\item{f_Cs_orig}{The ration of spending on the energy service to
the sum of initial spending on the energy service and other goods.}

\item{sigma}{The elasticity of substitution between spending on the energy service and spending on other goods.}

\item{graph_df_colnames}{A list of column names to use throughout the package.
Default is \code{ReboundTools::graph_df_colnames}.}
}
\value{
A version of \code{.DF} with new indifference curves added at the bottom.
}
\description{
Adds an indifference curve to a data frame of indifference curves
to be plotted on a preferences graph.
The indifference curves are accumulated in rows.
}
\details{
The preferences graph is \emph{always} indexed, so there is no \code{indexed} argument.
}
\examples{
meta <- tibble::tibble(Case = "Test case")
add_indifference_curve(meta = meta, line_name= "Test indifference curve",
                       qs1_qs0 = 2, Co1_Co0 = 3, f_Cs_orig = 0.0001, sigma = 0.2)
}
