% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{add_segment}
\alias{add_segment}
\title{Add a line segment to a data frame}
\usage{
add_segment(
  .DF = NULL,
  indexed,
  meta,
  graph_type,
  segment_name,
  colour = "black",
  size = 1,
  linetype = "solid",
  start_point = FALSE,
  end_arrow = FALSE,
  x_orig,
  y_orig,
  x,
  y,
  xend,
  yend,
  graph_df_colnames = ReboundTools::graph_df_colnames
)
}
\arguments{
\item{.DF}{A data frame that accumulates line segments.
When \code{NULL}, the default, a new data frame is created and returned.
When not \code{NULL}, rows for the segment are added to the bottom of \code{.DF}.}

\item{indexed}{A boolean telling whether the rebound path should be indexed to \code{1}
at its start.}

\item{meta}{A data frame of metadata for the segment to be added.
This metadata data frame provides the left-most columns of the return value.}

\item{graph_type}{The type of graph associated with this segment. See \code{ReboundTools::graph_types}.}

\item{segment_name}{A name for this segment.}

\item{colour}{The colour for this segment. Default is "black".}

\item{size}{The size (width) for this segment. Default is \code{1}.}

\item{linetype}{The line type for this segment. Default is "solid".}

\item{start_point}{A boolean that tells whether this segment would be plotted with a starting point. Default is \code{FALSE}.}

\item{end_arrow}{A boolean that tells whether this segment would be plotted with an ending arrow. Default is \code{FALSE}.}

\item{x_orig, y_orig}{The (x,y) coordinates of the starting point for this path,
used for indexing.}

\item{x, y}{The (x,y) coordinates of the starting point for this segment of the path.}

\item{xend, yend}{The (x,y) coordinates of the ending point for this segment of the path.}

\item{graph_df_colnames}{A list of column names to use throughout the package.
Default is \code{ReboundTools::graph_df_colnames}.}
}
\value{
A version of \code{.DF} with line segments added at the bottom.
}
\description{
Adds a line segment to a data frame of line segments.
The line segments are accumulated in rows.
}
\details{
There is usually no need to call this function directly.
Functions like \code{energy_paths()} call \code{add_segment()} internally.
}
\examples{
meta <- tibble::tibble(Case = "Test case")
add_segment(indexed = FALSE, meta = meta, graph_type = "Test type", 
  segment_name = "Test segment", 
  x_orig = 10, y_orig = 10, 
  x = 20, y = 30, xend = 40, yend = 50)
}
