% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_bar}
\alias{calc_bar}
\title{Calculate energy rebound data at the bar stage}
\usage{
calc_bar(
  .hat_data = NULL,
  tol = 1e-06,
  t_life_hat = ReboundTools::hat_vars$t_life_hat,
  R_alpha_hat = ReboundTools::hat_vars$R_alpha_hat,
  R_omega_hat = ReboundTools::hat_vars$R_omega_hat,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  p_E = ReboundTools::orig_vars$p_E,
  eta_engr_units_hat = ReboundTools::hat_vars$eta_engr_units_hat,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime,
  C_d_hat = ReboundTools::hat_vars$C_d_hat,
  C_dot_d_hat = ReboundTools::hat_vars$C_dot_d_hat,
  C_dot_om_hat = ReboundTools::hat_vars$C_dot_om_hat,
  C_dot_omd_hat = ReboundTools::hat_vars$C_dot_omd_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  e_qs_ps_UC_hat = ReboundTools::hat_vars$e_qs_ps_UC_hat,
  e_qo_ps_UC_hat = ReboundTools::hat_vars$e_qo_ps_UC_hat,
  e_qs_ps_C_hat = ReboundTools::hat_vars$e_qs_ps_C_hat,
  e_qo_ps_C_hat = ReboundTools::hat_vars$e_qo_ps_C_hat,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  t_life_bar = ReboundTools::bar_vars$t_life_bar,
  R_alpha_bar = ReboundTools::bar_vars$R_alpha_bar,
  R_omega_bar = ReboundTools::bar_vars$R_omega_bar,
  eta_engr_units_bar = ReboundTools::bar_vars$eta_engr_units_bar,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_d_bar = ReboundTools::bar_vars$C_d_bar,
  C_dot_d_bar = ReboundTools::bar_vars$C_dot_d_bar,
  C_dot_om_bar = ReboundTools::bar_vars$C_dot_om_bar,
  C_dot_omd_bar = ReboundTools::bar_vars$C_dot_omd_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  f_Cs_bar = ReboundTools::bar_vars$f_Cs_bar,
  e_qs_ps_UC_bar = ReboundTools::bar_vars$e_qs_ps_UC_bar,
  e_qo_ps_UC_bar = ReboundTools::bar_vars$e_qo_ps_UC_bar,
  e_qs_ps_C_bar = ReboundTools::bar_vars$e_qs_ps_C_bar,
  e_qo_ps_C_bar = ReboundTools::bar_vars$e_qo_ps_C_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar
)
}
\arguments{
\item{.hat_data}{An optional data frame containing rebound calculations, original data,
star data, and hat data,
likely calculated by \code{calc_hat()}.}

\item{tol}{The tolerance with which the budget constraint should be satisfied. Default is \code{1e-6}.}

\item{R_alpha_hat, R_omega_hat, eta_engr_units_hat, eta_hat, p_s_hat, C_dot_cap_hat, E_dot_emb_hat, M_dot_hat, q_dot_s_hat, N_dot_hat, M_dot_hat_prime, C_dot_o_hat, e_qs_ps_UC_hat, e_qo_ps_UC_hat, e_qs_ps_C_hat, e_qo_ps_C_hat, E_dot_s_hat}{See \code{ReboundTools::hat_vars}.}

\item{e_qs_M, e_qo_M, p_E}{See \code{ReboundTools::eeu_base_params}.}

\item{R_alpha_bar, R_omega_bar, eta_engr_units_bar, eta_bar, p_s_bar, C_dot_cap_bar, E_dot_emb_bar, M_dot_bar, q_dot_s_bar, E_dot_s_bar, C_dot_s_bar, C_dot_o_bar, f_Cs_bar, e_qs_ps_UC_bar, e_qo_ps_UC_bar, e_qs_ps_C_bar, e_qo_ps_C_bar, N_dot_bar}{See \code{ReboundTools::bar_vars}.}
}
\value{
A list or data frame of derived rebound values for the bar stage (after the income effect).
}
\description{
This function calculates energy rebound information for the bar
stage (immediately after the income effect).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat() \%>\% 
  calc_bar()
}
