% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_Deltas}
\alias{calc_Deltas}
\title{Calculate differences between stages}
\usage{
calc_Deltas(
  .tilde_data = NULL,
  rebound_vars = ReboundTools::rebound_vars,
  rebound_stages = ReboundTools::rebound_stages
)
}
\arguments{
\item{.tilde_data}{}

\item{rebound_vars}{}

\item{rebound_stages}{}
}
\value{

}
\description{
Calculate differences between stages
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat() \%>\% 
  calc_bar() \%>\% 
  calc_tilde() \%>\% 
  calc_Deltas()
}
