% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_hat}
\alias{calc_hat}
\title{Calculate energy rebound data at the hat stage}
\usage{
calc_hat(
  .star_data = NULL,
  eta_star = ReboundTools::star_vars$eta_star,
  p_s_star = ReboundTools::star_vars$p_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  e_qs_ps = ReboundTools::orig_vars$e_qs_ps,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  e_qo_ps = ReboundTools::orig_vars$e_qo_ps,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  C_dot_s_hat = ReboundTools::hat_vars$C_dot_s_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat
)
}
\arguments{
\item{.star_data}{An optional data frame containing EEU base data, original data,
and star data,
likely calculated by \code{calc_star()}.}
}
\value{
A list or data frame of derived rebound values for the hat stage (after the substitution effect).
}
\description{
This function calculates energy rebound information for the hat
stage (immediately after the substitution effect).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat()
}
