% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_hat}
\alias{calc_hat}
\title{Calculate energy rebound data at the hat stage}
\usage{
calc_hat(
  .star_data = NULL,
  use_sub_approx = FALSE,
  p_E = ReboundTools::orig_vars$p_E,
  e_qo_ps = ReboundTools::orig_vars$e_qo_ps,
  e_qs_ps_C = ReboundTools::orig_vars$e_qs_ps_C,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  sigma = ReboundTools::orig_vars$sigma,
  rho = ReboundTools::orig_vars$rho,
  eta_engr_units_star = ReboundTools::star_vars$eta_engr_units_star,
  eta_star = ReboundTools::star_vars$eta_star,
  p_s_star = ReboundTools::star_vars$p_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  M_dot_star = ReboundTools::star_vars$M_dot_star,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star,
  G_dot = ReboundTools::star_vars$G_dot,
  eta_engr_units_hat = ReboundTools::hat_vars$eta_engr_units_hat,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  C_dot_s_hat = ReboundTools::hat_vars$C_dot_s_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime
)
}
\arguments{
\item{.star_data}{An optional data frame containing EEU base data, original data,
and star data,
likely calculated by \code{calc_star()}.}

\item{use_sub_approx}{Tells whether to use an approximate method for
calculating the substitution effect.
Default is \code{FALSE}.
See details.}

\item{p_E}{See \code{ReboundTools::eeu_base_params}.}

\item{e_qo_ps, e_qs_ps_C, C_dot_cap_orig, C_dot_md_orig, f_Cs_orig, q_dot_s_orig, C_dot_o_orig, sigma, rho}{See \code{ReboundTools::orig_vars}.}

\item{eta_engr_units_star, eta_star, p_s_star, C_dot_cap_star, C_dot_md_star, E_dot_emb_star, M_dot_star, q_dot_s_star, eta_ratio, C_dot_o_star, N_dot_star, E_dot_s_star, G_dot}{See \code{ReboundTools::star_vars}.}

\item{eta_engr_units_hat, eta_hat, p_s_hat, C_dot_cap_hat, C_dot_md_hat, E_dot_emb_hat, M_dot_hat, q_dot_s_hat, E_dot_s_hat, C_dot_s_hat, C_dot_o_hat, N_dot_hat, M_dot_hat_prime}{See \code{ReboundTools::hat_vars}.}
}
\value{
A list or data frame of derived rebound values for the hat stage (after the substitution effect).
}
\description{
This function calculates energy rebound information for the hat
stage (immediately after the substitution effect).
}
\details{
By default, this function uses an exact method for calculating
the substitution effect.
The exact method uses a constant elasticity of substitution (CES)
formulation of utility,
which assumes constant elasticity of substitution (sigma) along
the indifference curve.

The approximate method is employed with
\code{use_sub_approx = TRUE}.
The approximate method
assumes constant price elasticity of energy service consumption.
(In fact, the price elasticity of energy service consumption
is not constant along an indifference curve.)
The approximation leads to a simple and elegant rebound expression
but a different (and incorrect) consumption bundle after the substitution effect.
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat()
}
