% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_orig}
\alias{calc_orig}
\title{Calculate energy rebound data at the original stage}
\usage{
calc_orig(
  .eeu_data = NULL,
  tol = 1e-06,
  r = ReboundTools::eeu_base_params$r,
  MJ_engr_unit = ReboundTools::eeu_base_params$MJ_engr_unit,
  p_E_engr_units = ReboundTools::eeu_base_params$p_E_engr_units,
  e_qs_ps_UC_orig = ReboundTools::eeu_base_params$e_qs_ps_UC_orig,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  eta_engr_units_orig = ReboundTools::orig_vars$eta_engr_units_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_cap_orig = ReboundTools::orig_vars$C_cap_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  C_dot_om_orig = ReboundTools::orig_vars$C_dot_om_orig,
  C_d_orig = ReboundTools::orig_vars$C_d_orig,
  E_emb_orig = ReboundTools::orig_vars$E_emb_orig,
  t_life_orig = ReboundTools::orig_vars$t_life_orig,
  R_alpha_orig = ReboundTools::orig_vars$R_alpha_orig,
  R_omega_orig = ReboundTools::orig_vars$R_omega_orig,
  p_E = ReboundTools::orig_vars$p_E,
  p_E_orig = ReboundTools::orig_vars$p_E_orig,
  eta_orig = ReboundTools::orig_vars$eta_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  R_alpha_C_dot_cap_orig = ReboundTools::orig_vars$R_alpha_C_dot_cap_orig,
  p_s_orig = ReboundTools::orig_vars$p_s_orig,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_d_orig = ReboundTools::orig_vars$C_dot_d_orig,
  R_omega_C_dot_d_orig = ReboundTools::orig_vars$R_omega_C_dot_d_orig,
  C_dot_omd_orig = ReboundTools::orig_vars$C_dot_omd_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  e_qo_ps_UC_orig = ReboundTools::orig_vars$e_qo_ps_UC_orig,
  e_qs_ps_C_orig = ReboundTools::orig_vars$e_qs_ps_C_orig,
  e_qo_ps_C_orig = ReboundTools::orig_vars$e_qo_ps_C_orig,
  sigma = ReboundTools::orig_vars$sigma,
  rho = ReboundTools::orig_vars$rho,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  N_dot_orig = ReboundTools::orig_vars$N_dot_orig
)
}
\arguments{
\item{.eeu_data}{An optional data frame containing EEU base data.
See \code{ReboundTools::eeu_base_params}.}

\item{tol}{The tolerance for checking internal consistency of rebound calculations. Default is \code{1e-6}.}

\item{r, MJ_engr_unit, p_E_engr_units, e_qs_ps_UC_orig, e_qs_M, e_qo_M}{See \code{ReboundTools::eeu_base_params}.}

\item{R_alpha_orig, R_omega_orig, eta_engr_units_orig, q_dot_s_orig, M_dot_orig, C_cap_orig, R_alpha_C_dot_cap_orig, C_d_orig, C_dot_omd_orig, C_dot_om_orig, C_dot_d_orig, R_omega_C_dot_d_orig, E_emb_orig, t_life_orig, p_E, eta_orig, E_dot_s_orig, C_dot_cap_orig, p_s_orig, C_dot_s_orig, C_dot_g_orig, f_Cs_orig, e_qo_ps_UC_orig, e_qs_ps_C_orig, e_qo_ps_C_orig, sigma, rho, E_dot_emb_orig, N_dot_orig, p_E_orig}{See \code{ReboundTools::orig_vars}.}
}
\value{
A list or data frame of derived rebound values.
}
\description{
This function calculates energy rebound information for the original
stage (pre-EEU).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig()
}
