% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_rebound}
\alias{calc_rebound}
\title{Calculate rebound terms.}
\usage{
calc_rebound(
  .Deltas_data = NULL,
  tol = 1e-06,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars[["∆E_dot_emb_star"]],
  Delta_C_dot_md_star = ReboundTools::Delta_vars[["∆C_dot_md_star"]],
  I_E = ReboundTools::eeu_base_params$I_E,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  e_qs_ps = ReboundTools::orig_vars$e_qs_ps,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars[["∆E_dot_s_hat"]],
  e_qo_ps = ReboundTools::orig_vars$e_qo_ps,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars[["∆C_dot_o_hat"]],
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  Delta_E_dot_s_bar = ReboundTools::Delta_vars[["∆E_dot_s_bar"]],
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars[["∆C_dot_o_bar"]],
  k = ReboundTools::eeu_base_params$k,
  Re_dempl = ReboundTools::rebound_terms$Re_dempl,
  Re_emb = ReboundTools::rebound_terms$Re_emb,
  Re_md = ReboundTools::rebound_terms$Re_md,
  Re_empl = ReboundTools::rebound_terms$Re_empl,
  Re_dsub = ReboundTools::rebound_terms$Re_dsub,
  Re_isub = ReboundTools::rebound_terms$Re_isub,
  Re_sub = ReboundTools::rebound_terms$Re_sub,
  Re_dinc = ReboundTools::rebound_terms$Re_dinc,
  Re_iinc = ReboundTools::rebound_terms$Re_iinc,
  Re_inc = ReboundTools::rebound_terms$Re_inc,
  Re_prod = ReboundTools::rebound_terms$Re_prod,
  Re_d = ReboundTools::rebound_terms$Re_d,
  Re_i = ReboundTools::rebound_terms$Re_i,
  Re_tot = ReboundTools::rebound_terms$Re_tot
)
}
\arguments{
\item{.Deltas_data}{}

\item{eta_ratio}{}

\item{e_qs_ps}{}
}
\value{

}
\description{
This function calculates rebound terms from a data frame that already contains ∆ terms.
Note that each rebound term is calculated twice as a way of validating the
derived expression for rebound.
}
