% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_rebound}
\alias{calc_rebound}
\title{Calculate rebound terms.}
\usage{
calc_rebound(
  .Deltas_data = NULL,
  tol = 1e-06,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars[["∆E_dot_emb_star"]],
  Delta_C_dot_md_star = ReboundTools::Delta_vars[["∆C_dot_md_star"]],
  I_E = ReboundTools::eeu_base_params$I_E,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  e_qs_ps = ReboundTools::orig_vars$e_qs_ps,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars[["∆E_dot_s_hat"]],
  e_qo_ps = ReboundTools::orig_vars$e_qo_ps,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars[["∆C_dot_o_hat"]],
  Re_dempl = ReboundTools::rebound_terms$Re_dempl,
  Re_emb = ReboundTools::rebound_terms$Re_emb,
  Re_md = ReboundTools::rebound_terms$Re_md,
  Re_dsub = ReboundTools::rebound_terms$Re_dsub,
  Re_isub = ReboundTools::rebound_terms$Re_isub
)
}
\arguments{
\item{.Deltas_data}{}

\item{eta_ratio}{}

\item{e_qs_ps}{}
}
\value{

}
\description{
This function calculates rebound terms from a data frame that already contains ∆ terms.
Note that each rebound term is calculated twice as a way of validating the
derived expression for rebound.
}
