% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_rebound_data}
\alias{calc_rebound_data}
\title{Calculate energy rebound derived data}
\usage{
calc_rebound_data(
  .eeu_data = NULL,
  MJ_engr_unit = ReboundTools::eeu_base_data$MJ_engr_unit,
  I_E = ReboundTools::eeu_base_data$I_E,
  k = ReboundTools::eeu_base_data$k,
  p_E = ReboundTools::eeu_base_data$p_E,
  eta_orig_engr_units = ReboundTools::eeu_base_data$eta_orig_engr_units,
  eta_tilde_engr_units = ReboundTools::eeu_base_data$eta_tilde_engr_units,
  e_ps_qs_UC = ReboundTools::eeu_base_data$e_ps_qs_UC,
  e_qs_M = ReboundTools::eeu_base_data$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_data$e_qo_M,
  q_dot_s_orig = ReboundTools::eeu_base_data$q_dot_s_orig,
  M_dot_orig = ReboundTools::eeu_base_data$M_dot_orig,
  C_cap_orig = ReboundTools::eeu_base_data$C_cap_orig,
  t_orig = ReboundTools::eeu_base_data$t_orig,
  C_cap_star = ReboundTools::eeu_base_data$C_cap_star,
  t_star = ReboundTools::eeu_base_data$t_star,
  C_dot_md_orig = ReboundTools::eeu_base_data$C_dot_md_orig,
  C_dot_md_star = ReboundTools::eeu_base_data$C_dot_md_star,
  E_emb_orig = ReboundTools::eeu_base_data$E_emb_orig,
  E_emb_star = ReboundTools::eeu_base_data$E_emb_star,
  eta_orig = ReboundTools::eeu_derived_data$eta_orig,
  eta_tilde = ReboundTools::eeu_derived_data$eta_tilde,
  eta_ratio = ReboundTools::eeu_derived_data$eta_ratio,
  E_dot_s_orig = ReboundTools::eeu_derived_data$E_dot_s_orig,
  S_dot_dev = ReboundTools::eeu_derived_data$S_dot_dev,
  G_dot = ReboundTools::eeu_derived_data$G_dot,
  C_dot_cap_orig = ReboundTools::eeu_derived_data$C_dot_cap_orig,
  C_dot_cap_star = ReboundTools::eeu_derived_data$C_dot_cap_star,
  p_s_orig = ReboundTools::eeu_derived_data$p_s_orig,
  p_s_star = ReboundTools::eeu_derived_data$p_s_star
)
}
\arguments{
\item{eta_orig, eta_tilde}{See \code{ReboundTools::eeu_base_data}.}

\item{.rebound_data}{An optional data frame containing EEU base data. See \code{ReboundTools::eeu_base_data}.}
}
\value{
A list or data frame of derived rebound values.
}
\description{
Calculate energy rebound derived data
}
\examples{
load_eeu_data() \%>\% 
  calc_rebound_data()
}
