% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_star}
\alias{calc_star}
\title{Calculate energy rebound data at the star stage}
\usage{
calc_star(
  .orig_data = NULL,
  MJ_engr_unit = ReboundTools::eeu_base_params$MJ_engr_unit,
  p_E = ReboundTools::orig_vars$p_E,
  eta_orig = ReboundTools::orig_vars$eta_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  eta_engr_units_star = ReboundTools::star_vars$eta_engr_units_star,
  E_emb_star = ReboundTools::star_vars$E_emb_star,
  t_life_star = ReboundTools::star_vars$t_life_star,
  C_cap_star = ReboundTools::star_vars$C_cap_star,
  t_own_star = ReboundTools::star_vars$t_own_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  eta_star = ReboundTools::star_vars$eta_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  G_dot = ReboundTools::star_vars$G_dot,
  p_s_star = ReboundTools::star_vars$p_s_star,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  C_dot_s_star = ReboundTools::star_vars$C_dot_s_star,
  M_dot_star = ReboundTools::star_vars$M_dot_star,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star
)
}
\arguments{
\item{.orig_data}{An optional data frame containing EEU base data and original data,
likely calculated by \code{calc_orig()}.}

\item{MJ_engr_unit, p_E}{See \code{ReboundTools::eeu_base_params}.}

\item{eta_orig, E_dot_s_orig, q_dot_s_orig, M_dot_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig}{See \code{ReboundTools::orig_vars}.}

\item{eta_engr_units_star, E_emb_star, t_life_star, C_cap_star, t_own_star, C_dot_md_star, eta_star, eta_ratio, S_dot_dev, G_dot, p_s_star, q_dot_s_star, C_dot_cap_star, E_dot_emb_star, C_dot_s_star, M_dot_star, N_dot_star, C_dot_o_star, E_dot_s_star}{See \code{ReboundTools::star_vars}.}
}
\value{
A list or data frame of derived rebound values for the star stage (after the emplacement effect).
}
\description{
This function calculates energy rebound information for the star
stage (immediately after the emplacement effect, with no behavior change).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star()
}
