% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rebound_data.R
\name{calc_tilde}
\alias{calc_tilde}
\title{Calculate energy rebound data at the tilde stage}
\usage{
calc_tilde(
  .bar_data = NULL,
  p_E = ReboundTools::orig_vars$p_E,
  t_life_bar = ReboundTools::bar_vars$t_life_bar,
  R_alpha_bar = ReboundTools::bar_vars$R_alpha_bar,
  R_omega_bar = ReboundTools::bar_vars$R_omega_bar,
  eta_engr_units_bar = ReboundTools::bar_vars$eta_engr_units_bar,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_dot_om_bar = ReboundTools::bar_vars$C_dot_om_bar,
  C_d_bar = ReboundTools::bar_vars$C_d_bar,
  C_dot_d_bar = ReboundTools::bar_vars$C_dot_d_bar,
  C_dot_omd_bar = ReboundTools::bar_vars$C_dot_omd_bar,
  C_dot_g_bar = ReboundTools::bar_vars$C_dot_g_bar,
  e_qs_ps_UC_bar = ReboundTools::bar_vars$e_qs_ps_UC_bar,
  e_qo_ps_UC_bar = ReboundTools::bar_vars$e_qo_ps_UC_bar,
  e_qs_ps_C_bar = ReboundTools::bar_vars$e_qs_ps_C_bar,
  e_qo_ps_C_bar = ReboundTools::bar_vars$e_qo_ps_C_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar,
  t_life_tilde = ReboundTools::tilde_vars$t_life_tilde,
  R_alpha_tilde = ReboundTools::tilde_vars$R_alpha_tilde,
  R_omega_tilde = ReboundTools::tilde_vars$R_omega_tilde,
  eta_engr_units_tilde = ReboundTools::tilde_vars$eta_engr_units_tilde,
  eta_tilde = ReboundTools::tilde_vars$eta_tilde,
  p_E_tilde = ReboundTools::tilde_vars$p_E_tilde,
  p_s_tilde = ReboundTools::tilde_vars$p_s_tilde,
  C_dot_cap_tilde = ReboundTools::tilde_vars$C_dot_cap_tilde,
  R_alpha_C_dot_cap_tilde = ReboundTools::tilde_vars$R_alpha_C_dot_cap_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  M_dot_tilde = ReboundTools::tilde_vars$M_dot_tilde,
  q_dot_s_tilde = ReboundTools::tilde_vars$q_dot_s_tilde,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  C_dot_s_tilde = ReboundTools::tilde_vars$C_dot_s_tilde,
  C_dot_om_tilde = ReboundTools::tilde_vars$C_dot_om_tilde,
  C_d_tilde = ReboundTools::tilde_vars$C_d_tilde,
  C_dot_d_tilde = ReboundTools::tilde_vars$C_dot_d_tilde,
  R_omega_C_dot_d_tilde = ReboundTools::tilde_vars$R_omega_C_dot_d_tilde,
  C_dot_omd_tilde = ReboundTools::tilde_vars$C_dot_omd_tilde,
  C_dot_g_tilde = ReboundTools::tilde_vars$C_dot_g_tilde,
  f_Cs_tilde = ReboundTools::tilde_vars$f_Cs_tilde,
  e_qs_ps_UC_tilde = ReboundTools::tilde_vars$e_qs_ps_UC_tilde,
  e_qo_ps_UC_tilde = ReboundTools::tilde_vars$e_qo_ps_UC_tilde,
  e_qs_ps_C_tilde = ReboundTools::tilde_vars$e_qs_ps_C_tilde,
  e_qo_ps_C_tilde = ReboundTools::tilde_vars$e_qo_ps_C_tilde,
  N_dot_tilde = ReboundTools::tilde_vars$N_dot_tilde
)
}
\arguments{
\item{.bar_data}{An optional data frame containing rebound calculations, original data,
star data, hat data, and bar data,
likely calculated by \code{calc_bar()}.}

\item{p_E}{See \code{ReboundTools::orig_vars}}

\item{t_life_bar, C_dot_om_bar, C_d_bar, C_dot_d_bar, C_dot_omd_bar, R_alpha_bar, R_omega_bar, eta_engr_units_bar, eta_bar, p_s_bar, C_dot_cap_bar, E_dot_emb_bar, M_dot_bar, q_dot_s_bar, E_dot_s_bar, C_dot_s_bar, C_dot_g_bar, e_qs_ps_UC_bar, e_qo_ps_UC_bar, e_qs_ps_C_bar, e_qo_ps_C_bar, N_dot_bar}{See \code{ReboundTools::bar_vars}.}

\item{t_life_tilde, C_dot_om_tilde, C_d_tilde, C_dot_d_tilde, R_omega_C_dot_d_tilde, C_dot_omd_tilde, R_alpha_tilde, R_omega_tilde, eta_engr_units_tilde, eta_tilde, p_s_tilde, C_dot_cap_tilde, R_alpha_C_dot_cap_tilde, E_dot_emb_tilde, M_dot_tilde, q_dot_s_tilde, E_dot_s_tilde, C_dot_s_tilde, C_dot_g_tilde, f_Cs_tilde, e_qs_ps_UC_tilde, e_qo_ps_UC_tilde, e_qs_ps_C_tilde, e_qo_ps_C_tilde, N_dot_tilde, p_E_tilde}{See \code{ReboundTools::tilde_vars}.}
}
\value{
A list or data frame of derived rebound values for the bar stage (after the income effect).
}
\description{
This function calculates energy rebound information for the tilde
stage (immediately after the macro effect).
}
\examples{
load_eeu_data() \%>\% 
  calc_orig() \%>\% 
  calc_star() \%>\% 
  calc_hat() \%>\% 
  calc_bar() \%>\% 
  calc_tilde()
}
