% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{cost_paths}
\alias{cost_paths}
\title{Make a data frame of segments
for a cost rebound graph.
Each stage of the rebound process is represented in the data frame.}
\usage{
cost_paths(
  .rebound_data,
  indexed = FALSE,
  graph_params = ReboundTools::default_graph_params,
  rebound_segments = ReboundTools::rebound_segments,
  graph_type = ReboundTools::graph_types$cost,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  Delta_C_dot_cap_star = ReboundTools::Delta_vars$Delta_C_dot_cap_star,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_C_dot_s_hat = ReboundTools::Delta_vars$Delta_C_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_C_dot_s_bar = ReboundTools::Delta_vars$Delta_C_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound analysis results,
likely created by \code{rebound_analysis()}.}

\item{indexed}{A boolean telling whether the rebound path should be indexed to \code{1}
at its start.}

\item{graph_params}{See \code{ReboundTools::graph_params}.}

\item{rebound_segments}{See \code{ReboundTools::rebound_segments}.}

\item{graph_type}{See \code{ReboundTools::graph_types}.}

\item{C_dot_s_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig}{See \code{ReboundTools::orig_vars}.}

\item{G_dot}{See \code{ReboundTools::star_vars}.}

\item{Delta_C_dot_cap_star, Delta_C_dot_md_star, Delta_C_dot_s_hat, Delta_C_dot_o_hat, Delta_C_dot_s_bar, Delta_C_dot_o_bar}{See \code{ReboundTools::Delta_vars}.}
}
\value{
A data frame with cost rebound path segments.
}
\description{
Make a data frame of segments
for a cost rebound graph.
Each stage of the rebound process is represented in the data frame.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  cost_paths()
}
