% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_graph_params}
\alias{default_graph_params}
\title{Graph parameters}
\format{
A string list with 70 entries.
\describe{

\item{lineend}{The line end style.}
\item{linejoin}{The line join style.}
\item{linejoin}{The line join style.}
\item{point_shape}{The shape for points between rebound effects. Default is \code{21}, a filled circle..}
\item{point_size}{The size for points between rebound effects. Default is \code{1}.}
\item{point_fill}{The fill colour for points between rebound effects. Default is "black".}
\item{point_stroke}{The size of the line surrounding points between rebound effects. Default is \code{1}.}
\item{point_colour}{The line colour for points between rebound effects. Default is "black".}
\item{dempl_colour}{The colour for direct emplacment lines.}
\item{emb_colour}{The colour for embodied energy lines.}
\item{cap_colour}{The colour for capital cost lines.}
\item{md_colour}{The colour for maintenance and disposal lines.}
\item{dsub_colour}{The colour for direct substitution lines.}
\item{isub_colour}{The colour for indirect substitution lines.}
\item{dinc_colour}{The colour for direct income lines.}
\item{iinc_colour}{The colour for indirect income lines.}
\item{prod_colour}{The colour for productivity lines.}
\item{dempl_size}{The size for direct emplacment lines.}
\item{emb_size}{The size for embodied energy lines.}
\item{cap_size}{The size for capital cost lines.}
\item{md_size}{The size for maintenance and disposal lines.}
\item{dsub_size}{The size for direct substitution lines.}
\item{isub_size}{The size for indirect substitution lines.}
\item{dinc_size}{The size for direct income lines.}
\item{iinc_size}{The size for indirect income lines.}
\item{prod_size}{The size for productivity lines.}
\item{dempl_linetyps}{The linetype for direct emplacment lines.}
\item{emb_linetype}{The linetype for embodied energy lines.}
\item{cap_linetype}{The linetype for capital cost lines.}
\item{md_linetype}{The linetype for maintenance and disposal lines.}
\item{dsub_linetype}{The linetype for direct substitution lines.}
\item{isub_linetype}{The linetype for indirect substitution lines.}
\item{dinc_linetype}{The linetype for direct income lines.}
\item{iinc_linetype}{The linetype for indirect income lines.}
\item{prod_linetype}{The linetype for productivity lines.}
\item{energy_grid_colour}{The colour for energy grid lines.}
\item{zero_perc_rebound_grid_colour}{The colour for energy the 0\% rebound lines.}
\item{hundred_perc_rebound_grid_colour}{The colour for the 100\% rebound lines.}
\item{energy_rebound_lines_colour}{The colour for energy rebound lines.}
\item{cost_grid_colour}{The colour for cost grid lines.}
\item{cost_ray_colour}{The colour for cost rays.}
\item{prefs_grid_colour}{The colour for preferences grid lines.}
\item{prefs_indiff_curve_colour}{The colour for indifference curve lines.}
\item{energy_grid_size}{The size for energy grid lines.}
\item{zero_perc_rebound_grid_size}{The size for 0\% rebound grid lines.}
\item{hundred_perc_rebound_grid_size}{The size for the 100\% rebound grid lines.}
\item{energy_rebound_lines_size}{The size for energy rebound lines.}
\item{cost_grid_size}{The size for cost grid lines.}
\item{cost_ray_size}{The size for cost rays.}
\item{prefs_grid_size}{The size for preferences grid lines.}
\item{prefs_indiff_curve_size}{The size for indifference curve lines.}
\item{energy_grid_linetype}{The linetype for energy grid lines.}
\item{zero_perc_rebound_grid_linetype}{The linetype for 0\% rebound grid lines.}
\item{hundred_perc_rebound_grid_linetype}{The linetype for the 100\% rebound grid lines.}
\item{energy_rebound_lines_linetype}{The linetype for energy rebound lines.}
\item{cost_grid_linetype}{The linetype for cost grid lines.}
\item{cost_ray_linetype}{The linetype for cost rays.}
\item{prefs_grid_linetype}{The linetype for preferences grid lines.}
\item{prefs_indiff_curve_linetype}{The linetype for indifference curve lines.}
\item{n_indiff_curve_points}{The number of points on the indifference curves.}
\item{qs_qs0_lower}{The lower bound for the x value in the indifference curves.}
\item{qs_qs0_upper}{The upper bound for the x value in the indifference curves.}
\item{include_start_point}{A boolean that tells whether to include the starting point.}
\item{start_point_size}{The size of the start point.}
\item{start_point_shape}{The shape of the start point.}
\item{include_end_arrow}{A boolean that tells whether to include an ending arrow.}
\item{arrow_angle}{The angle for the arrow head, in degrees.}
\item{arrow_length}{The length of the arrow. See \code{grid::unit}.}
\item{arrow_type}{The arrow type, "closed" (the default) or "open". See \code{grid::arrow}.}
}
}
\usage{
default_graph_params
}
\description{
The list of graph parameters for drawing
energy path graphs, cost path graphs, and preference path graphs
in the \code{ReboundTools} package.
}
\details{
This list is passed to several graphing functions.
Callers may pass a modified version of this list
to change graph appearance.
}
\examples{
default_graph_params
}
\keyword{datasets}
