% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeu_base_params}
\alias{eeu_base_params}
\title{EEU base parameters}
\format{
A string list with 13 entries.
\describe{
\item{reference}{A string for a reference for this case, e.g. bibliographic entry for a paper in which this case appears.}
\item{case}{A string to identify the case being analyzed, e.g., "Lamp".}
\item{original}{A string to identify the original device, e.g., "Incandescent".}
\item{upgrade}{A string to identify the upgraded device, e.g., "LED".}
\item{service_unit}{A string to identify the unit of the energy service, e.g., "miles" in "miles/gal" or "lm-hr" in "lm-hr/kW-hr".}
\item{energy_engr_unit}{A string to identify the energy units of the service, e.g., "gal" in "miles/gal" or "kW-hr" in "lm-hr/kW-hr".}
\item{MJ_engr_unit}{A unit conversion factor: the number of MJ per engineering unit for the service efficiency. For example, if the service efficiency is given in miles/gallon, \code{MJ_engr_unit} should be 126.6 MJ/gallon. This unit conversion number is used in calculating the actual service efficiency.}
\item{I_E}{The energy intensity of the economy [MJ/$].}
\item{k}{The macro effect factor [--].}
\item{p_E_engr_units}{The price of energy in engineering units, e.g., $/gal or $/kW-hr [$/energy_engr_unit].}
\item{e_qs_ps_UC_orig}{The original uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [--].}
\item{e_qs_M}{The income ("M") elasticity ("e") of energy service ("qs") consumption [--].}
\item{e_qo_M}{The income ("M") elasticity ("e") of other goods ("qo") consumption [--].}
}
}
\usage{
eeu_base_params
}
\description{
This is the list of the required parameters for a rebound analysis.
These parameters must be supplied for a rebound analysis to commence.
The names in this list are the assumed names throughout this package,
but they can always be overridden by function arguments.
}
\details{
In the descriptions for each parameter, typical units are provided for
energy efficiency upgrades for cars and electric lamps.

Note the following meanings for the default parameter names:
\describe{
\item{orig}{The original device, prior to the EEU.}
\item{star}{After the emplacement effect and before the substitution effect.}
\item{hat}{After the substitution effect and before the income effect.}
\item{bar}{After the income effect and before the macro effect.}
\item{tilde}{After the macro effect.}
\item{I}{An economic intensity (per $).}
\item{p}{Price.}
\item{E}{Energy, typically final energy.}
\item{s}{Energy service.}
\item{o}{Other goods.}
\item{eps}{Elasticity.}
\item{UC}{Uncompensated.}
\item{dot}{Signifies a rate, typically per year.}
\item{q}{Quantity of energy service or other goods consumed.}
\item{M}{Income.}
\item{C}{Cost.}
\item{t}{Expected device lifetime.}
}
}
\examples{
eeu_base_params
}
\keyword{datasets}
