% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeu_derived_data}
\alias{eeu_derived_data}
\title{EEU derived data}
\format{
A string list with 19 entries.
\describe{
\item{eta_ratio}{The ratio \code{eta_tilde/eta_orig}.}
}
}
\usage{
eeu_derived_data
}
\description{
This is the list of the derived parameters for a rebound analysis.
Each of these parameters is calculated data in \code{ReboundTools::eeu_base_data}.
}
\examples{
eeu_derived_data
}
\keyword{datasets}
