% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_df_colnames}
\alias{graph_df_colnames}
\title{Graph data frame columns}
\format{
A string list with 13 entries.
\describe{
\item{colour_col}{The name of the column containing line colours.}
\item{size_col}{The name of the column containing line sizes (widths).}
\item{linetype_col}{The name of the column containing line types.}
\item{graph_type_col}{The name of the column containing graph types.}
\item{line_name_col}{The name of the column containing names (string identifiers) for lines, segments, and grids.}
\item{slope_col}{The name of the column containing line slopes.}
\item{intercept_col}{The name of the column containing line intercepts.}
\item{x_col}{The name of the column containing starting x values.}
\item{y_col}{The name of the column containing starting y values.}
\item{xend_col}{The name of the column containing ending x values.}
\item{yend_col}{The name of the column containing ending y values.}
\item{start_point_col}{The name of the boolean column telling whether this row contains a segment that should have a starting point.}
\item{end_arrow_col}{The name of the boolean column telling whether this row contains a segment that should have an ending arrow.}
}
}
\usage{
graph_df_colnames
}
\description{
The list of names of graph data frame columns.
These are default names for columns produced internally.
}
\examples{
graph_df_colnames
}
\keyword{datasets}
