% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_df_colnames}
\alias{graph_df_colnames}
\title{Names of graph data frame columns}
\format{
A string list with 26 entries.
\describe{
\item{colour_col}{The name of the column containing line colours.}
\item{size_col}{The name of the column containing line sizes (widths).}
\item{linetype_col}{The name of the column containing line types.}
\item{graph_type_col}{The name of the column containing graph types.}
\item{line_name_col}{The name of the column containing names (string identifiers) for lines, segments, and grids.}
\item{slope_col}{The name of the column containing line slopes.}
\item{intercept_col}{The name of the column containing line intercepts.}
\item{x_col}{The name of the column containing starting x values.}
\item{y_col}{The name of the column containing starting y values.}
\item{xend_col}{The name of the column containing ending x values.}
\item{yend_col}{The name of the column containing ending y values.}
\item{qs1_qs0_col}{The name of the column containing a q_s/q_s_0 point on this indifference curve.}
\item{Cg1_Cg0_col}{The name of the column containing a C_g/C_g_0 point on this indifference curve.}
\item{f_Cs_orig_col}{The name of the column containing the original value of f_Cs for this indifference curve.}
\item{sigma_col}{The name of the column containing ending sigma values for this indifference curve.}
\item{start_point_col}{The name of the boolean column telling whether this row contains a segment that should have a starting point.}
\item{end_arrow_col}{The name of the boolean column telling whether this row contains a segment that should have an ending arrow.}
\item{Re_names}{The name of the column of rebound names.}
\item{Re_values}{The name of the column of rebound values.}
\item{y_names_col}{The name of the column of y-axis variable names.}
\item{y_vals_col}{The name of the column of y-axis variable names.}
}
}
\usage{
graph_df_colnames
}
\description{
The list of names of graph data frame columns.
These are default names for columns produced internally.
}
\examples{
graph_df_colnames
}
\keyword{datasets}
