% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_types}
\alias{graph_types}
\title{Graph types}
\format{
A string list with 3 entries.
\describe{
\item{energy}{The energy path graph type.}
\item{expenditure}{The expenditure path graph type.}
\item{consumption}{The consumption path graph type.}
}
}
\usage{
graph_types
}
\description{
The list of graph types in the \code{ReboundTools} package.
}
\examples{
graph_types
}
\keyword{datasets}
