% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hat_vars}
\alias{hat_vars}
\title{EEU hat data}
\format{
A string list with 5 entries.
\describe{
\item{eta_hat}{Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], exactly \code{eta_star}.}
\item{p_s_hat}{The energy service price after the substitution effect [$/service], exactly \code{p_s_star}.}
\item{C_dot_cap_hat}{The capital expenditure rate after the substitution effect [$/year], exactly \code{C_dot_cap_star}.}
\item{C_dot_md_hat}{The maintenance and disposal expenditure rate after the substitution effect [$/year], exactly \code{C_dot_md_star}.}
\item{q_dot_s_hat}{The rate of energy service consumption after the substitution effect[service/year], calculated by \code{q_dot_s_star * eta_ratio^(-e_qs_ps)}.}
}
}
\usage{
hat_vars
}
\description{
This is the list of the derived variables at the hat stage (after emplacement) of a rebound analysis.
}
\examples{
hat_vars
}
\keyword{datasets}
