% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{indifference_lines}
\alias{indifference_lines}
\title{Indifference curve lines}
\usage{
indifference_lines(
  .rebound_data,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  sigma = ReboundTools::orig_vars$sigma,
  prefs_type = ReboundTools::graph_types$preferences,
  grid_colour = ReboundTools::graph_colours$grid,
  grid_linetype = "solid",
  grid_size = 0.2
)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
This function calculates parameters for indifference curves
to be drawn on the preferences graph.
}
