% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indifference.R
\name{indifference_lines}
\alias{indifference_lines}
\title{Indifference curve lines}
\usage{
indifference_lines(
  .rebound_data,
  graph_type = ReboundTools::graph_types$preferences,
  graph_params = ReboundTools::default_graph_params,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  sigma = ReboundTools::orig_vars$sigma,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound information,
likely created by \code{rebound_analysis()}.}

\item{graph_type}{See \code{ReboundTools::graph_types}. Default is \code{ReboundTools::graph_types$preferences}.}

\item{graph_params}{Parameters that control the appearance of the graph. Default is \code{ReboundTools::default_graph_params}.}

\item{q_dot_s_orig, C_dot_o_orig, f_Cs_orig, sigma}{See \code{ReboundTools::orig_vars}.}

\item{q_dot_s_hat}{See \code{ReboundTools::hat_vars}.}

\item{q_dot_s_bar, C_dot_o_bar}{See \code{ReboundTools::bar_vars}.}

\item{x_vals}{The x-axis values at which the indifference curve is to be calculated.}

\item{q_dot_s_star, C_dot_o_star}{See \code{ReboundTools::star_vars}.}
}
\value{
A data frame of indifference curves for a preferences graph.
}
\description{
This function calculates indifference curves
to be drawn on the preferences graph.
}
\details{
The preferences graph is \emph{always} indexed, so there is no \code{indexed} argument.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  indifference_lines()
}
