% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{iso_budget_lines_prefs}
\alias{iso_budget_lines_prefs}
\title{Title}
\usage{
iso_budget_lines_prefs(
  .rebound_data,
  p_s_orig = ReboundTools::orig_vars$p_s_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  orig_size = 0.5,
  p_s_star = ReboundTools::star_vars$p_s_star,
  G_dot = ReboundTools::star_vars$G_dot,
  Delta_q_dot_s_hat = ReboundTools::Delta_vars$Delta_q_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  grid_colour = ReboundTools::graph_colours$orig
)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
Title
}
