% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lines.R
\name{iso_cost_lines}
\alias{iso_cost_lines}
\title{Create a data frame of iso-cost lines}
\usage{
iso_cost_lines(
  .rebound_data,
  indexed = FALSE,
  graph_type = ReboundTools::graph_types$cost,
  graph_params = ReboundTools::default_graph_params,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  G_dot = ReboundTools::star_vars$G_dot
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound information,
likely created by \code{rebound_analysis()}.}

\item{indexed}{A boolean telling whether the rebound path should be indexed to \code{1}
at its start.}

\item{graph_type}{See \code{ReboundTools::graph_types}.}

\item{graph_params}{See \code{ReboundTools::graph_params}.}

\item{C_dot_s_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig}{See \code{ReboundTools::orig_vars}.}

\item{G_dot}{See \code{ReboundTools::star_vars}.}
}
\value{
A data frame of iso-cost lines for a cost graph.
}
\description{
This function creates a data frame of constant-cost lines
at various rebound stages.
The iso lines are described by slope and intercept.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  iso_cost_lines()
}
