% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_analysis_vars}
\alias{key_analysis_vars}
\title{Key analysis variables}
\format{
A string list with 18 entries.
\describe{
\item{eta_engr_units}{Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service, in engineering units.}
\item{eta}{Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service.}
\item{p_s}{Energy service price [service/MJ], calculated by \code{p_E/eta}.}
\item{q_dot_s}{The rate of energy service consumption [service/yr], calculated by \code{eta*E_dot_s}.}
\item{E_dot_s}{The rate of final energy consumption by the energy conversion device [MJ/yr].}
\item{E_dot_emb}{The rate of embodied energy demand by the energy conversion device [MJ/yr], calculated by \code{E_emb/t_life}.}
\item{C_dot_s}{The expenditure rate of energy consumption by the device [$/yr], calculated by \code{p_s*q_dot_s}.}
\item{C_dot_cap}{The capital expenditure rate of the device [$/yr], calculated by \code{C_cap/t_own}.}
\item{C_dot_md}{The maintenance and disposal expenditure rate of the device [$/yr].}
\item{C_dot_o}{The other goods consumption rate [$/yr], calculated, initially, as a residual of the budget constraint.}
\item{N_dot}{Net income [$/yr].}
\item{M_dot}{Real income [$/yr].}
}
}
\usage{
key_analysis_vars
}
\description{
This is the list of key rebound analysis variables.
These variables are differenced later in the analysis.
}
\examples{
key_analysis_vars
}
\keyword{datasets}
