% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orig_vars}
\alias{orig_vars}
\title{EEU orig data}
\format{
A string list with 28 entries.
\describe{
\item{R_alpha_orig}{The original discount multiplier for beginning-of-life expenses, such as capital cost.}
\item{R_omega_orig}{The original discount multiplier for end-of-life expenses, such as disposal cost.}
\item{p_E}{The price of energy [$/MJ], calculated by \code{p_E_engr_units / MJ_engr_unit}.}
\item{q_dot_s_orig}{The original (pre-EEU) consumption rate of the energy service. Example units are [miles/yr] [lumen-hours/yr].}
\item{C_cap_orig}{The net capital expenditure of the original device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].}
\item{M_dot_orig}{The disposable income rate, exclusive of taxes and savings [$/yr].}
\item{t_own_orig}{The expected ownership duration of the original device [yr].}
\item{C_dot_md_orig}{The original (pre-EEU) maintenance and disposal expenditure rate [$/yr].}
\item{eta_engr_units_orig}{The original (pre-EEU) energy service efficiency.  This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of \code{eta_engr_units_orig} is assumed to be the same as the denominator unit of \code{MJ_engr_unit}.}
\item{eta_orig}{Energy service efficiency of the original (pre-EEU) device on a per-MJ basis [service/MJ], calculated by \code{eta_engr_units_orig / MJ_engr_unit}.}
\item{E_dot_s_orig}{The final energy consumption rate of the original (pre-EEU) device [MJ/yr], calculated by \code{q_dot_s_orig / eta_orig}.}
\item{C_dot_cap_orig}{The original (pre-EEU) capital expenditure rate [$/yr], calculated by \code{C_cap_orig / t_orig}.}
\item{p_s_orig}{The original (pre-EEU) energy service price [$/service], calculated by \code{p_E / eta_orig}.}
\item{C_dot_s_orig}{The original (pre-EEU) rate of energy expenditures for the device [$/yr], calculated by \code{p_E * E_dot_s_orig}.}
\item{C_dot_o_orig}{The original (pre-EEU) rate of expenditure on other goods [$/yr], calculated by \code{M_dot_orig - C_dot_s_orig - C_dot_cap_orig - C_dot_md_orig}.}
\item{f_Cs_orig}{The original (pre-EEU) fraction of the energy and other budget spent on the energy service [--], calculated by \code{C_dot_s_orig / (C_dot_s_orig + C_dot_o_orig)}.}
\item{e_qs_ps_C_orig}{The original compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [--], calculated by \code{e_qs_ps_UC_orig + f_Cs_orig*e_qs_M}.}
\item{e_qo_ps_C_orig}{The original compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [--], calculated by \code{f_Cs_orig*(f_Cs_orig + e_qs_ps_UC_orig) / (f_Cs_orig - 1)}.}
\item{e_qo_ps_UC_orig}{The original uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of other goods ("qo") consumption (cross-price elasticity) [--].}
\item{sigma}{The elasticity of substitution between energy service consumption and other goods consumption [--].}
\item{rho}{The exponent in the CES utility model, defined as rho = 1/sigma - 1 [--].}
\item{E_emb_orig}{The embodied energy of the original (pre-EEU) device [MJ].}
\item{t_life_orig}{The expected lifetime of the original (pre-EEU) device [yr].}
\item{E_dot_emb_orig}{The original (pre-EEU) rate of embodied energy demand[MJ/yr], calculated by \code{E_dot_emb / t_orig}.}
\item{N_dot_orig}{The original (pre-EEU) freed cash rate [$/yr], exactly \code{0}.}
}
}
\usage{
orig_vars
}
\description{
This is the list of the derived variables at the original (pre-EEU) stage of a rebound analysis.
}
\examples{
orig_vars
}
\keyword{datasets}
