% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{parametric_studies}
\alias{parametric_studies}
\title{Run parametric analyses}
\usage{
parametric_studies(original_cases, parameterization)
}
\arguments{
\item{original_cases}{A data frame with baseline data from which parametric studies are desired.}

\item{parameterization}{A data frame with sweeps for variables in \code{original_cases}.
\code{expand.grid()} will be called on all columns of \code{parameterization},
Case\verb{r eeu_base_params$case}.}
}
\value{
A data frame containing parametric studies.
}
\description{
After defining a base case for rebound analyses,
parametric analyses and sensitivity studies are
often helpful.
This function provides a way to do parametric studies.
}
\details{
\code{original_cases} must have the same structure as the example data frames in this package.
See the file at \code{sample_eeu_data_path()}.

\code{parameterization} must be a data frame with column ``
}
\examples{

}
