% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{prefs_paths}
\alias{prefs_paths}
\title{A data frame of preferences paths}
\usage{
prefs_paths(
  .rebound_data,
  graph_params = ReboundTools::default_graph_params,
  rebound_segments = ReboundTools::rebound_segments,
  graph_type = ReboundTools::graph_types$preferences,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  Delta_q_dot_s_hat = ReboundTools::Delta_vars$Delta_q_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_q_dot_s_bar = ReboundTools::Delta_vars$Delta_q_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  graph_df_colnames = ReboundTools::graph_df_colnames
)
}
\arguments{
\item{.rebound_data}{A data frame of rebound analysis results,
likely created by \code{rebound_analysis()}.}

\item{graph_params}{See \code{ReboundTools::graph_params}.}

\item{rebound_segments}{See \code{ReboundTools::rebound_segments}.}

\item{graph_type}{See \code{ReboundTools::graph_types}.}

\item{q_dot_s_star, C_dot_o_star}{See \code{ReboundTools::star_vars}.}

\item{Delta_q_dot_s_hat, Delta_C_dot_o_hat, Delta_q_dot_s_bar, Delta_C_dot_o_bar}{See \code{ReboundTools::Delta_vars}.}

\item{graph_df_colnames}{See \code{ReboundTools::graph_df_colnames}.}
}
\value{
A data frame of information for creating preference graphs.
}
\description{
Make a data frame of segments
for a preferences graph.
Each stage of the rebound process is represented in the data frame.
}
\details{
The preferences graph is \emph{always} indexed, so there is no \code{indexed} argument.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  prefs_paths()
}
