% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_graphs}
\alias{rebound_graphs}
\title{Create a path map for rebound analysis}
\usage{
rebound_graphs(.path_data)
}
\arguments{
\item{.path_data}{The paths to be graphed}
}
\value{
A ggplot2 object with the graphs
}
\description{
Create a path map for rebound analysis
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  rebound_paths() \%>\% 
  rebound_graphs()
}
