% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_graphs}
\alias{rebound_graphs}
\title{Create rebound graphs}
\usage{
rebound_graphs(
  .rebound_data,
  indexed = FALSE,
  energy_path_graph = TRUE,
  cost_path_graph = TRUE,
  preferences_path_graph = TRUE,
  graph_params = ReboundTools::default_graph_params
)
}
\arguments{
\item{.rebound_data}{Rebound data, likely loaded by \code{load_eeu_data()}.}
}
\value{
A \code{ggplot2} object with the graphs
}
\description{
Create rebound graphs
}
\examples{
load_eeu_data() \%>\% 
  rebound_graphs()
}
