% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_graphs_helper}
\alias{rebound_graphs_helper}
\title{Create path maps for rebound analysis}
\usage{
rebound_graphs_helper(
  .path_data,
  .grid_data = NULL,
  .indifference_data = NULL,
  graph_df_colnames = ReboundTools::graph_df_colnames
)
}
\arguments{
\item{.path_data}{A data frame of paths to be added to the graph.
The columns "colour" and "size" control the colour and width of the segment}

\item{.grid_data}{A data frame of lines to be added to the graph.}

\item{.indifference_data}{A data frame of indifference curves to be added to the graph.}

\item{graph_df_colnames}{The names of column names in data frames of graph data. Default is \code{ReboundTools::graph_df_colnames}.}
}
\value{
A \code{ggplot2} object with the graphs
}
\description{
This is a helper function for \code{rebound_graphs()}.
There is normally no need to call this function.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  energy_paths() \%>\% 
  rebound_graphs_helper()
}
