% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_paths}
\alias{rebound_paths}
\title{Title}
\usage{
rebound_paths(
  .rebound_data,
  indexed = FALSE,
  I_E = ReboundTools::eeu_base_params$I_E,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  S_dot_dev_colour = "red",
  S_dot_dev_size = 0.1,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  G_dot_colour = "red",
  G_dot_size = 1,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars$Delta_E_dot_emb_star,
  Delta_E_dot_emb_star_colour = "red",
  Delta_E_dot_emb_star_size = 1,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_C_dot_md_star_I_E_colour = "black",
  Delta_C_dot_md_star_I_E_size = 0.5,
  Delta_C_dot_md_star_colour = "red",
  Delta_C_dot_md_star_size = 1,
  Delta_C_dot_cap_star = ReboundTools::Delta_vars$Delta_C_dot_cap_star,
  Delta_C_dot_cap_star_colour = "black",
  Delta_C_dot_cap_star_size = 0.5,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars$Delta_E_dot_s_hat,
  Delta_E_dot_s_hat_colour = "orange",
  Delta_E_dot_s_hat_size = 1,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_C_dot_o_hat_I_E_colour = "orange",
  Delta_C_dot_o_hat_I_E_size = 1,
  Delta_E_dot_s_bar = ReboundTools::Delta_vars$Delta_E_dot_s_bar,
  Delta_E_dot_s_bar_colour = "darkgreen",
  Delta_E_dot_s_bar_size = 1,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  Delta_C_dot_o_bar_I_E_colour = "darkgreen",
  Delta_C_dot_o_bar_I_E_size = 1,
  energy_type = "Energy",
  cost_type = "Cost",
  prefs_type = "Preferences"
)
}
\arguments{
\item{.rebound_data}{}
}
\value{

}
\description{
Title
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  rebound_paths()
}
