% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rebound_results_table}
\alias{rebound_results_table}
\title{Build a rebound results table}
\usage{
rebound_results_table(
  .analysis_data,
  escape_latex = TRUE,
  include_subtotals = TRUE,
  include_total = TRUE,
  as_percent = TRUE,
  rebound_terms = ReboundTools::rebound_terms,
  latex_rebound_terms = ReboundTools::latex_rebound_terms,
  case = ReboundTools::eeu_base_params$case,
  subtotals = c(ReboundTools::rebound_terms$Re_om, ReboundTools::rebound_terms$Re_d,
    ReboundTools::rebound_terms$Re_empl, ReboundTools::rebound_terms$Re_sub,
    ReboundTools::rebound_terms$Re_inc, ReboundTools::rebound_terms$Re_dir,
    ReboundTools::rebound_terms$Re_indir, ReboundTools::rebound_terms$Re_micro),
  total = ReboundTools::rebound_terms$Re_tot,
  term_name = "Rebound term",
  latex_term_name = "LaTeX rebound term",
  Re_val_colname = "Value [-]",
  perc_Re_val_colname = "Value [\%]",
  latex_Re_val_colname = "Value [--]",
  latex_perc_Re_val_colname = "Value [\\\\\%]",
  ...
)
}
\arguments{
\item{.analysis_data}{Rebound analysis results. Probably the output of a call to \code{rebound_analysis()}.}

\item{escape_latex}{When \code{TRUE} (the default), return LaTeX-compatible versions of strings.}

\item{include_subtotals}{Tells whether to include rebound subtotals
by stage and direct/indirect.
Default is \code{TRUE}.}

\item{include_total}{Tells whether to include rebound total.
Default is \code{TRUE}.}

\item{as_percent}{When \code{TRUE}, rebound results are reported as percentages.
When \code{FALSE}, rebound results are reported as fractions.
Default is \code{TRUE}.}

\item{rebound_terms}{See \code{ReboundTools::rebound_terms}.}

\item{latex_rebound_terms}{See \code{ReboundTools::latex_rebound_terms}.}

\item{case}{See \code{ReboundTools::eeu_base_params}.}

\item{subtotals}{The rebound terms that represent subtotals.
Default is \code{c(ReboundTools::rebound_terms$Re_empl, ReboundTools::rebound_terms$Re_sub, ReboundTools::rebound_terms$Re_inc, ReboundTools::rebound_terms$Re_dir, ReboundTools::rebound_terms$Re_indir, ReboundTools::rebound_terms$Re_micro)}.}

\item{total}{The rebound term that represents total rebound. Default is \code{ReboundTools::rebound_terms$Re_tot}.}

\item{term_name}{The title of the rebound term column. Default is "Rebound term".}

\item{latex_term_name}{The LaTeX term name column. Default is "LaTeX rebound term".}

\item{Re_val_colname}{The title of the rebound value column. Default is "Value [--]".}

\item{perc_Re_val_colname}{The title of the rebound value column when percentages are requested. Default is "Value [\%]".}

\item{latex_Re_val_colname}{The LaTeX rebound value column name. Default is "Value [--]".}

\item{latex_perc_Re_val_colname}{The LaTeX rebound value column name when percentages are requested. Default is "Value [\\\%]".}

\item{...}{Arguments passed to \code{xtable::xtable()}, possibly
\code{label}, \code{caption}, \code{digits}, etc.}
}
\value{
An \code{xtable} object suitable for printing.
}
\description{
Build a rebound results table
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  rebound_results_table()
}
