% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebound_terms}
\alias{rebound_terms}
\title{Rebound terms}
\format{
A string list with 18 entries.
\describe{
\item{Re_dempl}{Direct emplacement effect rebound, always 0.}
\item{Re_emb}{Indirect embodied energy effect rebound.}
\item{Re_cap}{Indirect capital expenditure effect rebound.}
\item{Re_om}{Indirect operations and maintenance expenditure effect rebound.}
\item{Re_d}{Indirect disposal effect rebound.}
\item{Re_omd}{Indirect operations, maintenance, and disposal effect rebound.}
\item{Re_empl}{Emplacement effect rebound.}
\item{Re_dsub}{Direct substitution effect rebound.}
\item{Re_isub}{Indirect substitution effect rebound.}
\item{Re_sub}{Substitution effect rebound.}
\item{Re_dinc}{Direct income effect rebound.}
\item{Re_iinc}{Indirect income effect rebound.}
\item{Re_inc}{Income effect rebound.}
\item{Re_micro}{Sum of all micro rebound effects.}
\item{Re_macro}{Indirect macro effect rebound.}
\item{Re_dir}{Sum of all direct rebound effects.}
\item{Re_indir}{Sum of all indirect rebound effects.}
\item{Re_tot}{Total rebound.}
}
}
\usage{
rebound_terms
}
\description{
This is the list of the rebound terms.
}
\examples{
rebound_terms
}
\keyword{datasets}
