% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rebound_terms_graph}
\alias{rebound_terms_graph}
\title{A sensitivity graph containing all rebound terms}
\usage{
rebound_terms_graph(
  .parametric_data = parametric_analysis(rebound_data, parameterization),
  rebound_data,
  parameterization,
  x_var,
  Re_terms = unlist(ReboundTools::rebound_terms),
  line_var = y_names_col,
  include_x_axis = TRUE,
  y_names_col = ReboundTools::graph_df_colnames$y_names_col,
  y_vals_col = ReboundTools::graph_df_colnames$y_vals_col,
  graph_params = ReboundTools::sens_graph_params,
  point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
  sweep_points = ReboundTools::parametric_analysis_point_types$sweep,
  orig_points = ReboundTools::parametric_analysis_point_types$orig
)
}
\arguments{
\item{.parametric_data}{A data frame, likely the result of calling \code{parametric_analysis()}.
Default is \code{parametric_analysis(rebound_data, parameterization)}.}

\item{rebound_data}{Rebound data, likely read by \code{load_eeu_data()}.
Default is \code{NULL}.}

\item{parameterization}{A list of lists that gives parameter sweeps.
At the top level, the list items must be named for cases in \code{rebound_data}.
At the next level, the parameters to be swept should be given,
along with their sweep values.
See examples.
Default is \code{NULL}.}

\item{x_var}{Strings that identify the x-axis and y-axis variables for this sensitivity graph.
These variables must appear in \code{.parametric_data}.
\code{x_var} must be a single string.
\code{y_var} can be a vector of strings.
See examples.}

\item{Re_terms}{A string vector that tells which rebound terms to include in the graph.
Default is \code{unlist(ReboundTools::rebound_terms)}.}

\item{line_var}{The name of variable to be used to discriminate lines on the graph.
Default is \code{y_names_col}.}

\item{include_x_axis}{A boolean that tells whether to include a horizontal line at y = 0.
Default is \code{TRUE}.}

\item{y_names_col, y_vals_col}{See \code{ReboundTools::graph_df_colnames}.}

\item{graph_params}{A list of parameters to control graph appearance.
See \code{ReboundTools::sens_graph_params}.}

\item{point_type_colname, sweep_points, orig_points}{See \code{ReboundTools::parametric_analysis_point_types}.}
}
\value{
A ggplot2 graph showing sensitivity of all rebound terms to \code{x_var}.
}
\description{
Create a sensitivity graph with lines for all rebound terms.
}
\details{
This function has the same arguments as \code{sensitivity_graphs()},
except that \code{yvar} is missing.
\code{y_var} is missing, because the ordinate is assumed to be rebound values,
and all rebound components are included.
}
\examples{
df <- load_eeu_data()
sens_params <- list(Car = list(eta_engr_units_star = seq(35, 50, by = 0.5)), 
                    Lamp = list(eta_engr_units_star = seq(70, 90, by = 5)))
rebound_terms_graph(rebound_data = df, parameterization = sens_params, 
                    x_var = "eta_engr_units_tilde") +
  ggplot2::facet_wrap(facets = "Case", scales = "free_x")
}
