% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{sensitivity_graphs}
\alias{sensitivity_graphs}
\title{Sensitivity graphs for rebound analyses}
\usage{
sensitivity_graphs(
  .parametric_data = parametric_analysis(rebound_data, parameterization),
  rebound_data,
  parameterization,
  x_var,
  y_vars = ReboundTools::rebound_terms,
  linetype_var = ReboundTools::eeu_base_params$case,
  linecolour_var = ReboundTools::eeu_base_params$case,
  linesize_var = ReboundTools::eeu_base_params$case,
  graph_params = ReboundTools::sens_graph_params,
  point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
  sweep_points = ReboundTools::parametric_analysis_point_types$sweep,
  orig_points = ReboundTools::parametric_analysis_point_types$orig,
  Re_names = ReboundTools::graph_df_colnames$Re_names,
  Re_values = ReboundTools::graph_df_colnames$Re_values
)
}
\arguments{
\item{.parametric_data}{A data frame, likely the result of calling \code{parametric_analysis()}.
Default is \code{parametric_analysis(rebound_data, parameterization)}.}

\item{rebound_data}{Rebound data, likely read by \code{load_eeu_data()}.}

\item{parameterization}{A named list of lists. See details.}

\item{x_var, y_var}{Strings that identify the x-axis and y-axis variables for this sensitivity graph.
These variables must appear in \code{.parametric_data}.}

\item{linetype_var, linecolour_var, linesize_var}{Strings that identify variables to be used for
type, color, and size of lines.
Default is \code{ReboundTools::eeu_base_params$case}.}

\item{graph_params}{A list of parameters to control graph appearance.
See \code{ReboundTools::sens_graph_params}.}

\item{point_type_colname, sweep_points, orig_points}{See \code{ReboundTools::parametric_analysis_point_types}.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function makes sensitivity graphs for rebound analysis.
}
\details{
The caller can adjust the aesthetics of the graph with manual scales.
}
\examples{
# Sensitivity of total rebound (Re_tot) to productivity multiplier (k)
df <- load_eeu_data()
sens_params <- list(Car = list(k = seq(0.5, 1.5, by = 0.5)), 
                    Lamp = list(k = seq(0, 2, by = 1)))
sensitivity_graphs(rebound_data = df, parameterization = sens_params, 
                   x_var = "k", y_var = "Re_tot") +
 ggplot2::scale_colour_manual(values = c(Car = "black", Lamp = "black")) + 
 ggplot2::scale_size_manual(values = c(Car = 0.5, Lamp = 0.5)) + 
 ggplot2::scale_linetype_manual(values = c(Car = "solid", Lamp = "dashed"))
# A more-complicated example that shows multi-variate sensitivity.
# Values of the productivity parameter (k) is shown in rows of the lattice plot.
# Uncompensated price elasticity of energy service consumption (e_qs_ps_UC) 
# is shown in columns of the lattice plot.
# Total rebound (Re_tot) is given on the y-axis, and 
# energy intensity of the economy (I_E) is given on the x-axis.
# The cases (Car and Lamp) are shown as different lines.
sens_params_2 <- list(Car = list(k = seq(0, 2, by = 0.5), 
                               I_E = seq(2, 5, by = 1), 
                               e_qs_ps_UC = seq(-0.5, -0.1, by = 0.1)), 
                    Lamp = list(k = seq(0, 2, by = 0.5),
                                I_E = seq(2, 5, by = 1), 
                                e_qs_ps_UC = seq(-0.5, -0.1, by = 0.1)))
sensitivity_graphs(rebound_data = df, parameterization = sens_params_2, 
                   x_var = "I_E", y_var = "Re_tot") +
  ggplot2::facet_grid(rows = ggplot2::vars(k), 
                      cols = ggplot2::vars(e_qs_ps_UC), scales = "free_y") +
  ggplot2::scale_colour_manual(values = c(Car = "darkgreen", Lamp = "black")) + 
  ggplot2::scale_size_manual(values = c(Car = 0.5, Lamp = 1)) + 
  ggplot2::scale_linetype_manual(values = c(Car = "solid", Lamp = "dotted")) + 
  ggplot2::labs(colour = ggplot2::element_blank(), 
                size = ggplot2::element_blank(),
                linetype = ggplot2::element_blank())
}
