% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{stages_table}
\alias{stages_table}
\title{Build a stages table}
\usage{
stages_table(
  .analysis_data,
  add_units = TRUE,
  escape_latex = TRUE,
  vars = ReboundTools::key_analysis_vars,
  latex_vars = ReboundTools::latex_key_analysis_vars,
  stages = ReboundTools::rebound_stages,
  latex_stages = ReboundTools::latex_rebound_stages,
  case = ReboundTools::eeu_base_params$case,
  service_unit = ReboundTools::eeu_base_params$service_unit,
  energy_engr_unit = ReboundTools::eeu_base_params$energy_engr_unit,
  ...,
  .var = ".var",
  .stage = ".stage",
  .var_stage = ".var_stage",
  .value = ".value",
  .name = ".name",
  .unit_col = ".unit_col"
)
}
\arguments{
\item{.analysis_data}{A data frame, usually the result of calling \code{rebound_analysis()}. Default is \code{rebound_analysis(load_eeu_data(file))}.}

\item{add_units}{When \code{TRUE} (the default), adds a unit specification to variable names in the table.}

\item{escape_latex}{When \code{TRUE} (the default), return LaTeX-compatible versions of strings.}

\item{vars}{A list of variables for rows of the table. Default is \code{ReboundTools::key_analysis_vars}.
Variable order is preserved in the table.}

\item{latex_vars}{See \code{ReboundTools::latex_key_analysis_vars}. Set \code{NULL} to prevent conversion to LaTeX variable names.}

\item{stages}{A list of stages for columns of the table. Default is \code{ReboundTools::rebound_stages}.
Stage order is preserved in the table.}

\item{latex_stages}{See \code{ReboundTools::latex_rebound_stages}. Set \code{NULL} to prevent conversion to LaTeX stage names.}

\item{case}{See \code{ReboundTools::eeu_base_params}.}

\item{service_unit, energy_engr_unit}{See \code{ReboundTools::eeu_base_params}.}

\item{...}{Arguments passed to \code{xtable::xtable()}, possibly
\code{label}, \code{caption}, \code{digits}, etc.}

\item{.var, .stage, .var_stage, .value, .name, .unit_col}{Column names used internally.}
}
\value{
An \code{xtable} object suitable for printing.
}
\description{
A stages table has variables in rows and stages in columns.
These tables enable tracking of variables across the different stages of rebound.
}
\examples{
load_eeu_data() \%>\% 
  rebound_analysis() \%>\% 
  stages_table()
}
