% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{stages_table}
\alias{stages_table}
\title{Build a stages table}
\usage{
stages_table(
  .analysis_data = rebound_analysis(load_eeu_data(file)),
  add_units = TRUE,
  escape_latex = TRUE,
  vars = ReboundTools::key_analysis_vars,
  latex_vars = ReboundTools::latex_key_analysis_vars,
  stages = ReboundTools::rebound_stages,
  latex_stages = ReboundTools::latex_rebound_stages,
  file = sample_eeu_data_path(),
  case = ReboundTools::eeu_base_params$case,
  service_unit = ReboundTools::eeu_base_params$service_unit,
  energy_engr_unit = ReboundTools::eeu_base_params$energy_engr_unit,
  ...
)
}
\arguments{
\item{add_units}{When \code{TRUE} (the default), adds a unit specification to variable names in the table.}

\item{vars}{A list of variables for rows of the table. Default is \code{ReboundTools::key_analysis_vars}.
Variable order is preserved in the table.}

\item{latex_vars}{See \code{ReboundTools::latex_key_analysis_vars}. Set \code{NULL} to prevent conversion to LaTeX variable names.}

\item{stages}{A list of stages for columns of the table. Default is \code{ReboundTools::rebound_stages}.
Stage order is preserved in the table.}

\item{latex_stages}{See \code{ReboundTools::latex_rebound_stages}. Set \code{NULL} to prevent conversion to LaTeX stage names.}

\item{file}{An optional path to a file. Default is \code{sample_eeu_data_path()}.}

\item{case}{See \code{ReboundTools::eeu_base_params}.}

\item{...}{Arguments passed to \code{xtable::xtable()}, possibly
\code{label}, \code{caption}, \code{digits}, etc.}

\item{.results}{A data frame, usually the result of calling \code{rebound_analysis()}. Default is \code{rebound_analysis(load_eeu_data(file))}.}
}
\value{
An \code{xtable} object giving the details of the table.
}
\description{
A stages table has variables in rows and stages in columns.
These tables enable tracking of variables across the different stages of rebound.
}
\examples{
stages_table()
}
