% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\title{Units for variables}
\usage{
units(
  .var_name,
  service_unit,
  energy_engr_unit,
  escape_latex = FALSE,
  surround_left = "[",
  surround_right = "]",
  leading_delta = "^Delta_",
  energy_si = "MJ",
  time_unit = "year",
  currency = "$",
  currency_latex = "\\\\$",
  unitless = "-",
  unitless_latex = "--",
  p_E_engr_units = "p_E_engr_units",
  p_E = "p_E",
  p_s = "p_s_",
  q_dot_s = "q_dot_s_",
  efficiency_engr_units = "eta_engr_units",
  efficiency = "eta",
  k = "k",
  I_E = "I_E",
  elasticities = "e_",
  time = "t_",
  cost_rate = "C_dot_",
  cost = "C_",
  income_rate = "M_dot_",
  income = "M_",
  freed_cash_rate = "N_dot_",
  freed_cash = "N_",
  energy_rate = "E_dot",
  energy = "E",
  S_dot_dev = "S_dot_dev",
  G_dot = "G_dot",
  rebound = "Re_",
  f_Cs_ = "f_Cs_"
)
}
\arguments{
\item{.var_name}{}

\item{service_unit}{}

\item{energy_engr_unit}{}

\item{leading_delta}{}

\item{currency}{}

\item{time}{}

\item{trailing_stage}{}
}
\value{

}
\description{
Units for variables
}
